% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{add_chronology}
\alias{add_chronology}
\title{Add a new chronology to a collection unit.}
\usage{
add_chronology(object, x, y)
}
\arguments{
\item{object}{A collectionunit object}

\item{x}{A chronology object}

\item{y}{A \code{data.frame} of sample ages}
}
\value{
chronology object defined by user,
}
\description{
Add a new chronology to a collection unit.
}
