% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{as.data.frame,author-method}
\alias{as.data.frame,author-method}
\title{Convert a publication author to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{author}(x)
}
\arguments{
\item{x}{An author}
}
\description{
Convert a publication author to a \code{data.frame}
}
