% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{as.data.frame,authors-method}
\alias{as.data.frame,authors-method}
\title{Convert a publication author to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{authors}(x)
}
\arguments{
\item{x}{An author}
}
\value{
\code{data.frame} with publications metadata
}
\description{
Convert a publication author to a \code{data.frame}
}
