% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{as.data.frame,chronologies-method}
\alias{as.data.frame,chronologies-method}
\title{as.data.frame datasets}
\usage{
\S4method{as.data.frame}{chronologies}(x)
}
\arguments{
\item{x}{chronologies object}
}
\description{
show as dataframe as prep to save as csv
}
