% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{as.data.frame,chronology-method}
\alias{as.data.frame,chronology-method}
\title{as.data.frame chronology}
\usage{
\S4method{as.data.frame}{chronology}(x)
}
\arguments{
\item{x}{chronology object}
}
\description{
show as dataframe as prep to save as csv
}
