% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{as.data.frame,collunit-method}
\alias{as.data.frame,collunit-method}
\title{as.data.frame site}
\usage{
\S4method{as.data.frame}{collunit}(x)
}
\arguments{
\item{x}{site object}
}
\value{
data.frame object with a collection units metadata
}
\description{
show as dataframe
}
