% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{as.data.frame,collunits-method}
\alias{as.data.frame,collunits-method}
\title{as.data.frame collunits}
\usage{
\S4method{as.data.frame}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\value{
data.frame of multiple collection units metadata.
}
\description{
show as dataframe as prep to save as csv
}
