% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{as.data.frame,contact-method}
\alias{as.data.frame,contact-method}
\title{Transform a contacts object to a data.frame()}
\usage{
\S4method{as.data.frame}{contact}(x)
}
\arguments{
\item{x}{A contact object.}
}
\value{
\code{data.frame} object with contact metadata
}
\description{
Transform a contacts object to a data.frame()
}
