% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{as.data.frame,contacts-method}
\alias{as.data.frame,contacts-method}
\title{Transform a contacts object to a data.frame()}
\usage{
\S4method{as.data.frame}{contacts}(x)
}
\arguments{
\item{x}{A contacts object.}
}
\value{
\code{data.frame} object with multiple \code{contacts} metadata
}
\description{
Transform a contacts object to a data.frame()
}
