% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{as.data.frame,dataset-method}
\alias{as.data.frame,dataset-method}
\title{as.data.frame dataset}
\usage{
\S4method{as.data.frame}{dataset}(x)
}
\arguments{
\item{x}{dataset object}
}
\value{
\code{data.frame} with \code{dataset} metadata
}
\description{
show as dataframe as prep to save as csv
}
