% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdataset-methods.R
\name{as.data.frame,datasets-method}
\alias{as.data.frame,datasets-method}
\title{as.data.frame datasets}
\usage{
\S4method{as.data.frame}{datasets}(x)
}
\arguments{
\item{x}{datasets object}
}
\description{
show as dataframe as prep to save as csv
}
