% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{as.data.frame,publication-method}
\alias{as.data.frame,publication-method}
\title{Convert a \code{publication} to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{publication}(x)
}
\arguments{
\item{x}{A \code{publication} object.}
}
\value{
\code{data.frame} with publications' metadata.
}
\description{
Convert a \code{publication} to a \code{data.frame}
}
