% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{as.data.frame,publications-method}
\alias{as.data.frame,publications-method}
\title{Convert publications to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{publications}(x)
}
\arguments{
\item{x}{A \code{publications} object.}
}
\value{
\code{data.frame} with publications' metadata.
}
\description{
Convert publications to a \code{data.frame}
}
