% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{as.data.frame,sample-method}
\alias{as.data.frame,sample-method}
\title{as.data.frame sample}
\usage{
\S4method{as.data.frame}{sample}(x)
}
\arguments{
\item{x}{sample object}
}
\description{
show as dataframe
}
