% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{as.data.frame,site-method}
\alias{as.data.frame,site-method}
\title{as.data.frame site}
\usage{
\S4method{as.data.frame}{site}(x)
}
\arguments{
\item{x}{site object}
}
\value{
\code{data.frame} object with site metadata
}
\description{
show as dataframe as prep to save as csv
}
