% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{as.data.frame,sites-method}
\alias{as.data.frame,sites-method}
\title{as.data.frame sites}
\usage{
\S4method{as.data.frame}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\description{
show as dataframe as prep to save as csv
}
