% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{as.data.frame,specimen-method}
\alias{as.data.frame,specimen-method}
\title{as.data.frame specimen}
\usage{
\S4method{as.data.frame}{specimen}(x)
}
\arguments{
\item{x}{specimen object}
}
\value{
\code{data.frame} with \verb{specimen metadata}
}
\description{
show as data.frame
}
