% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{as.data.frame,speleothem-method}
\alias{as.data.frame,speleothem-method}
\title{as.data.frame speleothem}
\usage{
\S4method{as.data.frame}{speleothem}(x)
}
\arguments{
\item{x}{speleothem object}
}
\value{
\code{data.frame} with \code{speleothem} metadata
}
\description{
show as dataframe as prep to save as csv
}
