% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{as.data.frame,speleothems-method}
\alias{as.data.frame,speleothems-method}
\title{as.data.frame speleothems}
\usage{
\S4method{as.data.frame}{speleothems}(x)
}
\arguments{
\item{x}{speleothems object}
}
\value{
\code{data.frame} with \code{speleothems} metadata
}
\description{
show as dataframe as prep to save as csv
}
