% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{as.data.frame,taxa-method}
\alias{as.data.frame,taxa-method}
\title{as.data.frame taxa}
\usage{
\S4method{as.data.frame}{taxa}(x)
}
\arguments{
\item{x}{taxa object}
}
\value{
\code{data.frame} with \code{taxa} metadata
}
\description{
show as dataframe as prep to save as csv
}
