% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{as.data.frame,taxon-method}
\alias{as.data.frame,taxon-method}
\title{as.data.frame taxon}
\usage{
\S4method{as.data.frame}{taxon}(x)
}
\arguments{
\item{x}{taxon object}
}
\value{
\code{data.frame} with \code{taxon} metadata
}
\description{
show as dataframe as prep to save as csv
}
