% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/contact-methods.R, R/dataset-methods.R, R/publication-methods.R,
%   R/site-methods.R, R/speleothem-methods.R, R/taxon-methods.R
\name{as.data.frame,chronology-method}
\alias{as.data.frame,chronology-method}
\alias{as.data.frame,chronologies-method}
\alias{as.data.frame,collunit-method}
\alias{as.data.frame,collunits-method}
\alias{as.data.frame,contact-method}
\alias{as.data.frame,contacts-method}
\alias{as.data.frame,dataset-method}
\alias{as.data.frame,datasets-method}
\alias{as.data.frame,authors-method}
\alias{as.data.frame,publication-method}
\alias{as.data.frame,publications-method}
\alias{as.data.frame}
\alias{as.data.frame,sites-method}
\alias{as.data.frame,speleothem-method}
\alias{as.data.frame,speleothems-method}
\alias{as.data.frame,taxon-method}
\alias{as.data.frame,taxa-method}
\title{as.data.frame}
\usage{
\S4method{as.data.frame}{chronology}(x)

\S4method{as.data.frame}{chronologies}(x)

\S4method{as.data.frame}{collunit}(x)

\S4method{as.data.frame}{collunits}(x)

\S4method{as.data.frame}{contact}(x)

\S4method{as.data.frame}{contacts}(x)

\S4method{as.data.frame}{dataset}(x)

\S4method{as.data.frame}{datasets}(x)

\S4method{as.data.frame}{authors}(x)

\S4method{as.data.frame}{publication}(x)

\S4method{as.data.frame}{publications}(x)

\S4method{as.data.frame}{site}(x)

\S4method{as.data.frame}{sites}(x)

\S4method{as.data.frame}{speleothem}(x)

\S4method{as.data.frame}{speleothems}(x)

\S4method{as.data.frame}{taxon}(x)

\S4method{as.data.frame}{taxa}(x)
}
\arguments{
\item{x}{\code{neotoma2} object}
}
\value{
\code{data.frame} object
}
\description{
Returns \code{neotoma2} object's data as a \code{data.frame}.
}
