% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{as.list,sites-method}
\alias{as.list,sites-method}
\title{as.list sites}
\usage{
\S4method{as.list}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
\code{list} object with sites metadata
}
\description{
show as dataframe as prep to save as csv
}
