% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{authors-class}
\alias{authors-class}
\title{An S4 class for a set of Neotoma author objects.}
\description{
The S4 \code{authors} are a set of individual \code{author} objects that
are then associated with a single S4 \code{publication} class.
}
