% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_chron.R
\name{build_chron}
\alias{build_chron}
\title{get_downloads}
\usage{
build_chron(x)
}
\arguments{
\item{x}{dataset list}
}
\value{
list parsed into datasets
}
\description{
Helper function to build a dataset
}
\examples{
\dontrun{
# To build dataset from API call:
build_dataset(x)
}

}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
