% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collunit.R
\name{build_collunit}
\alias{build_collunit}
\title{Build a collection unit from the API response}
\usage{
build_collunit(x)
}
\arguments{
\item{x}{The structured JSON from a Neotoma API v2.0 response that returns a collection unit in any form.}
}
\value{
An object of class \code{collunit}
}
\description{
Build a collection unit from the API response
}
