% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_collunit.R
\name{build_collunits}
\alias{build_collunits}
\title{Build a collection unit from the API response}
\usage{
build_collunits(x)
}
\arguments{
\item{x}{The structured JSON from a Neotoma API v2.0 response that
returns a collection unit in any form.}
}
\value{
An simple \code{collunit} object
}
\description{
Build a collection unit from the API response
}
