% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sample.R
\name{build_sample}
\alias{build_sample}
\title{Build a samples \code{data.frame} from Neotoma API JSON}
\usage{
build_sample(x)
}
\arguments{
\item{x}{sample list}
}
\value{
A simple \code{sample} object
}
\description{
Helper function to build a sample from the API input (list formatted)
coming from the Neotoma API.
}
\author{
Socorro Dominguez \email{s.dominguez@ht-data.com}
}
