% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_sites}
\alias{build_sites}
\title{Build a \code{site} from the Neotoma API response.}
\usage{
build_sites(x)
}
\arguments{
\item{x}{A list returned from the Neotoma API's \code{data} slot.}
}
\description{
Build a \code{site} from the Neotoma API response.
}
\examples{
\dontrun{
response <- jsonlite::fromJSON('https://api.neotomadb.org/v2.0/data/datasets/100,101',
                               flatten=FALSE, simplifyVector=FALSE)
response <- cleanNULL(response)
newSites <- build_sites(response$data)
newSites
}
}
