% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_specimen.R
\name{build_specimen}
\alias{build_specimen}
\title{Build specimen objects.}
\usage{
build_specimen(x)
}
\arguments{
\item{x}{specimen list}
}
\value{
A list (from JSON) parsed into specimen
}
\description{
A helper function to build a specimen object from a list returned
within the get_speciments() call.
}
\examples{
\dontrun{
build_specimen(x)
}

}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
