% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{c,chronologies-method}
\alias{c,chronologies-method}
\title{c Method - Combine chronologies objects}
\usage{
\S4method{c}{chronologies}(x, y)
}
\arguments{
\item{x}{chronologies object 1}

\item{y}{chronologies object 2}
}
\value{
concatenated \code{chronologies}
}
\description{
c Method - Combine chronologies objects
}
