% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{c,collunits-method}
\alias{c,collunits-method}
\title{c Method - Combine collunits objects}
\usage{
\S4method{c}{collunits}(x, y)
}
\arguments{
\item{x}{collunits object 1}

\item{y}{collunits object 2}
}
\value{
concatenated collection units without duplicates
}
\description{
c Method - Combine collunits objects
}
