% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{c,contacts-method}
\alias{c,contacts-method}
\title{c Method - Combine contacts objects}
\usage{
\S4method{c}{contacts}(x, y)
}
\arguments{
\item{x}{contacts object 1}

\item{y}{contacts object 2}
}
\value{
concatenated and clean objects
}
\description{
c Method - Combine contacts objects
}
