% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{c,datasets-method}
\alias{c,datasets-method}
\title{c Method - Combine datasets objects}
\usage{
\S4method{c}{datasets}(x, y)
}
\arguments{
\item{x}{datasets object 1}

\item{y}{datasets object 2}
}
\value{
concatenated \code{datasets} object
}
\description{
c Method - Combine datasets objects
}
