% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{c,missingOrNULL-method}
\alias{c,missingOrNULL-method}
\title{c Method for NULL values}
\usage{
\S4method{c}{missingOrNULL}(x = "missingORNULL", y)
}
\arguments{
\item{x}{NULL object}

\item{y}{sites/datasets object}
}
\description{
c Method for NULL values
}
