% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{c,publications-method}
\alias{c,publications-method}
\title{Combine publication objects.}
\usage{
\S4method{c}{publications}(x, y)
}
\arguments{
\item{x}{A publications object.}

\item{y}{A publications object}
}
\value{
concatenated \code{publications} object
}
\description{
Combine publication objects.
}
