% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{c,samples-method}
\alias{c,samples-method}
\title{c Method - Combine datasets objects}
\usage{
\S4method{c}{samples}(x, y)
}
\arguments{
\item{x}{datasets object 1}

\item{y}{datasets object 2}
}
\value{
concatenated \code{samples} object
}
\description{
c Method - Combine datasets objects
}
