% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{c,sites-method}
\alias{c,sites-method}
\title{c Method - Combine sites objects}
\usage{
\S4method{c}{sites}(x, y)
}
\arguments{
\item{x}{sites object 1}

\item{y}{sites object 2}
}
\value{
concatenated and cleaned \code{sites} object
}
\description{
c Method - Combine sites objects
}
