% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{c,specimens-method}
\alias{c,specimens-method}
\title{c Method - Combine specimens objects}
\usage{
\S4method{c}{specimens}(x, y)
}
\arguments{
\item{x}{specimens object 1}

\item{y}{specimens object 2}
}
\value{
concatenated \code{specimens} object
}
\description{
c Method - Combine specimens objects
}
