% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{c,speleothems-method}
\alias{c,speleothems-method}
\title{c Method - Combine speleothems objects}
\usage{
\S4method{c}{speleothems}(x, y)
}
\arguments{
\item{x}{speleothems object 1}

\item{y}{speleothems object 2}
}
\value{
concatenated \code{speleothems} object
}
\description{
c Method - Combine speleothems objects
}
