% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{c,taxa-method}
\alias{c,taxa-method}
\title{c Method - Combine taxa objects}
\usage{
\S4method{c}{taxa}(x, y)
}
\arguments{
\item{x}{taxa object 1}

\item{y}{taxa object 2}
}
\value{
concatenated \code{taxa} object
}
\description{
c Method - Combine taxa objects
}
