% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{$,collunit-method}
\alias{$,collunit-method}
\title{$}
\usage{
\S4method{$}{collunit}(x, name)
}
\arguments{
\item{x}{collunit object}
}
\description{
Obtain slots of a collunit without using at-mark
}
