% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{$,contact-method}
\alias{$,contact-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{$}{contact}(x, name)
}
\arguments{
\item{x}{A contact object.}

\item{name}{The name of a contact slot.}
}
\value{
value in the selected slot
}
\description{
Extract or Replace Parts of an Object
}
