% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{$,contacts-method}
\alias{$,contacts-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{$}{contacts}(x, name)
}
\arguments{
\item{x}{A contacts object.}

\item{name}{The name of a contacts slot.}
}
\value{
object value from the slot
}
\description{
Extract or Replace Parts of an Object
}
