% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{$,dataset-method}
\alias{$,dataset-method}
\title{$}
\usage{
\S4method{$}{dataset}(x, name)
}
\arguments{
\item{x}{dataset object}

\item{name}{name of the slot}
}
\value{
Obtain a \code{dataset}'s \code{slot} value using $
}
\description{
Obtain slots of a dataset without using at-mark
}
