% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdataset-methods.R
\name{$,datasets-method}
\alias{$,datasets-method}
\title{$ for datasets}
\usage{
\S4method{$}{datasets}(x, name)
}
\arguments{
\item{x}{datasets object}
}
\description{
Obtain slots of a site without using at-mark
}
