% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{$,publication-method}
\alias{$,publication-method}
\title{Extract an element from a \code{publication}}
\usage{
\S4method{$}{publication}(x, name)
}
\arguments{
\item{x}{A \code{publication} object.}

\item{name}{The slot to obtain (e.g., \code{articletitle})}
}
\value{
\code{value} in the selected slot
}
\description{
Extract an element from a \code{publication}
}
