% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/asamples-methods.R
=======
% Please edit documentation in R/samples-methods.R
>>>>>>> dev
\name{$,sample-method}
\alias{$,sample-method}
\title{$}
\usage{
\S4method{$}{sample}(x, name)
}
\arguments{
\item{x}{sample object}
}
\description{
Obtain slots of a sample without using at-mark
}
