% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$<-,chronology-method}
\alias{$<-,chronology-method}
\title{$ Assignment}
\usage{
\S4method{$}{chronology}(x, name) <- value
}
\arguments{
\item{x}{chronology object}

\item{name}{name of the slot}
}
\description{
Obtain slots of a chronology without using at-mark
}
