% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{$<-,site-method}
\alias{$<-,site-method}
\title{Assign site field by numeric index}
\usage{
\S4method{$}{site}(x, name) <- value
}
\arguments{
\item{x}{The site object.}

\item{name}{name of the slot}

\item{value}{The value to be used.}
}
\value{
\code{site} object with reassigned values
}
\description{
Assign site field by numeric index
}
