% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{$<-,specimen-method}
\alias{$<-,specimen-method}
\title{Assign specimen field by numeric index}
\usage{
\S4method{$}{specimen}(x, name) <- value
}
\arguments{
\item{x}{The specimen object.}

\item{name}{name of the slot.}

\item{value}{The value to be used.}
}
\value{
\code{specimen} object with reassigned values
}
\description{
Assign specimen field by numeric index
}
