% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{$<-,speleothem-method}
\alias{$<-,speleothem-method}
\title{Assign speleothem field by numeric index}
\usage{
\S4method{$}{speleothem}(x, name) <- value
}
\arguments{
\item{x}{The speleothem object.}

\item{name}{name of the slot.}

\item{value}{The value to be used.}
}
\value{
Assign new \code{speleothem} by numeric index
}
\description{
Assign speleothem field by numeric index
}
