% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{$<-,taxon-method}
\alias{$<-,taxon-method}
\title{Assign taxon field by numeric index}
\usage{
\S4method{$}{taxon}(x, name) <- value
}
\arguments{
\item{x}{The taxon object.}

\item{name}{name of the slot.}

\item{value}{The value to be used.}
}
\value{
Assign new \code{taxon} by numeric index
}
\description{
Assign taxon field by numeric index
}
