% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{$,site-method}
\alias{$,site-method}
\title{$}
\usage{
\S4method{$}{site}(x, name)
}
\arguments{
\item{x}{site object}

\item{name}{name of the slot}
}
\value{
value at chosen slot in the \code{site} object
}
\description{
Obtain slots of a site without using at-mark
}
