% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{$,sites-method}
\alias{$,sites-method}
\title{$ for sites}
\usage{
\S4method{$}{sites}(x, name)
}
\arguments{
\item{x}{sites object}

\item{name}{name of the slot}
}
\value{
value at chosen slot in the \code{site} object
}
\description{
Obtain slots of a site without using at-mark
}
