% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{$,specimens-method}
\alias{$,specimens-method}
\title{$ for specimens}
\usage{
\S4method{$}{specimens}(x, name)
}
\arguments{
\item{x}{specimens object}

\item{name}{name of the slot.}
}
\value{
value at chosen slot in the \code{site} object
}
\description{
Obtain slots of a specimen without using at-mark
}
