% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{$,speleothem-method}
\alias{$,speleothem-method}
\title{$}
\usage{
\S4method{$}{speleothem}(x, name)
}
\arguments{
\item{x}{speleothem object}

\item{name}{name of the slot}
}
\value{
Obtain a \code{speleothem}'s \code{slot} value using $
}
\description{
Obtain slots of a speleothem without using at-mark
}
