% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{$,speleothems-method}
\alias{$,speleothems-method}
\title{$ for speleothems}
\usage{
\S4method{$}{speleothems}(x, name)
}
\arguments{
\item{x}{speleothems object}

\item{name}{name of the slot.}
}
\value{
Obtain a \code{speleothems}' \code{slot} value using $
}
\description{
Obtain slots of a speleothem without using at-mark
}
