% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{$,taxa-method}
\alias{$,taxa-method}
\title{$ for taxa}
\usage{
\S4method{$}{taxa}(x, name)
}
\arguments{
\item{x}{taxa object}

\item{name}{name of the slot.}
}
\value{
Obtain a \code{taxa}' \code{slot} value using $
}
\description{
Obtain slots of a taxon without using at-mark
}
