% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{$,taxon-method}
\alias{$,taxon-method}
\title{$}
\usage{
\S4method{$}{taxon}(x, name)
}
\arguments{
\item{x}{taxon object}

\item{name}{name of the slot}
}
\value{
Obtain a \code{taxon}'s \code{slot} value using $
}
\description{
Obtain slots of a taxon without using at-mark
}
