% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/contact-methods.R, R/dataset-methods.R, R/publication-methods.R,
%   R/samples-methods.R, R/site-methods.R, R/speleothem-methods.R,
%   R/taxon-methods.R
\name{$,chronology-method}
\alias{$,chronology-method}
\alias{$,chronologies-method}
\alias{$,collunit-method}
\alias{$,collunits-method}
\alias{$,contact-method}
\alias{$,contacts-method}
\alias{$,dataset-method}
\alias{$,datasets-method}
\alias{$,publication-method}
\alias{$,samples-method}
\alias{$,sample-method}
\alias{cash}
\alias{$,sites-method}
\alias{$,speleothem-method}
\alias{$,speleothems-method}
\alias{$,taxon-method}
\alias{$,taxa-method}
\title{cash}
\usage{
\S4method{$}{chronology}(x, name)

\S4method{$}{chronologies}(x, name)

\S4method{$}{collunit}(x, name)

\S4method{$}{collunits}(x, name)

\S4method{$}{contact}(x, name)

\S4method{$}{contacts}(x, name)

\S4method{$}{dataset}(x, name)

\S4method{$}{datasets}(x, name)

\S4method{$}{publication}(x, name)

\S4method{$}{samples}(x, name)

\S4method{$}{sample}(x, name)

\S4method{$}{site}(x, name)

\S4method{$}{sites}(x, name)

\S4method{$}{speleothem}(x, name)

\S4method{$}{speleothems}(x, name)

\S4method{$}{taxon}(x, name)

\S4method{$}{taxa}(x, name)
}
\arguments{
\item{x}{\code{neotoma2} object}

\item{name}{name of the slot}
}
\value{
value at chosen slot in the \code{site} object
}
\description{
Obtain slots of a site without using at-mark
}
