% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contacts.R
\name{check_contacts.contacts}
\alias{check_contacts.contacts}
\title{Get contact information for Neotoma contributors}
\usage{
\method{check_contacts}{contacts}(x, similarity = 0.5, ...)
}
\description{
Get contact information for Neotoma contributors
}
