% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contacts.R
\name{check_contacts}
\alias{check_contacts}
\title{Check contact information for a record against Neotoma contributors}
\usage{
check_contacts(x, ...)
}
\arguments{
\item{x}{contacts A \code{contacts} object associated with a set of names.}

\item{...}{Additional parameters associated with the call.}
}
\value{
contacts object
}
\description{
Uses the Neotoma API to search and access information about individuals
who have contributed to the data in the Neotoma Paleoecology Database
}
\author{
Simon Goring \email{goring@wisc.edu}
}
