% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,chronologies-method}
\alias{chroncontrols,chronologies-method}
\title{Recover the chron controls for a set of chronologies.}
\usage{
\S4method{chroncontrols}{chronologies}(x)
}
\arguments{
\item{x}{chronologies object}
}
\value{
data.frame with chronological controls
}
\description{
For a \code{chronologies} object return the chronological controls
used in building each chronology, labelled by chronology.
}
