% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,chronology-method}
\alias{chroncontrols,chronology-method}
\title{Recover the chron controls for a single chronology.}
\usage{
\S4method{chroncontrols}{chronology}(x)
}
\arguments{
\item{x}{chronology object}
}
\value{
data.frame with chronological controls
}
\description{
For a \code{chronology} object return the chronological controls
used in building it.
}
