% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,site-method}
\alias{chroncontrols,site-method}
\title{chroncontrols}
\usage{
\S4method{chroncontrols}{site}(x)
}
\arguments{
\item{x}{site object}
}
\description{
Obtain elements on the samples level
}
