% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,sites-method}
\alias{chroncontrols,sites-method}
\title{chroncontrols}
\usage{
\S4method{chroncontrols}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\description{
Obtain chroncontrols
}
