% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{chroncontrols}
\alias{chroncontrols}
\title{chroncontrols}
\usage{
chroncontrols(x)
}
\arguments{
\item{x}{Sites object to extract chroncontrols table from}
}
\description{
Show the samples table
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
