% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{chronologies-class}
\alias{chronologies-class}
\alias{chronologies}
\title{S4 class for chronologies information}
\description{
The grouped class for datasets
from the Neotoma Paleoecology Database.
}
