% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{chronologies,collunit-method}
\alias{chronologies,collunit-method}
\title{Extract chronologies from a sites object.}
\usage{
\S4method{chronologies}{collunit}(x)
}
\arguments{
\item{object}{A collunit object}
}
\description{
Extract chronologies from a sites object.
}
