% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{chronologies,collunits-method}
\alias{chronologies,collunits-method}
\title{Extract \code{chronologies} from a \code{collunits} object.}
\usage{
\S4method{chronologies}{collunits}(x)
}
\arguments{
\item{x}{A \code{collunits} object}
}
\value{
\code{chronologies} from a \code{collunits} object
}
\description{
Extract \code{chronologies} from a \code{collunits} object.
}
