% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{chronologies,site-method}
\alias{chronologies,site-method}
\title{Extract \code{chronologies} from a \code{site} object.}
\usage{
\S4method{chronologies}{site}(x)
}
\arguments{
\item{x}{A \code{site} object}
}
\value{
\code{chronologies} from a \code{site} object
}
\description{
Extract \code{chronologies} from a \code{site} object.
}
