% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{chronologies,sites-method}
\alias{chronologies,sites-method}
\title{Extract \code{chronologies} from a \code{sites} object.}
\usage{
\S4method{chronologies}{sites}(x)
}
\arguments{
\item{x}{A \code{sites} object}
}
\value{
\code{chronologies} from a \code{sites} object
}
\description{
Extract \code{chronologies} from a \code{sites} object.
}
