% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R,
%   R/02_genericDefinitions.R, R/subset-methods.R
\docType{class}
\name{chronologies}
\alias{chronologies}
\alias{chronologies-class}
\alias{chronologies,collunits-method}
\alias{chronologies,site-method}
\alias{chronologies,sites-method}
\title{S4 class for \code{chronologies} information}
\usage{
chronologies(x)

\S4method{chronologies}{collunits}(x)

\S4method{chronologies}{site}(x)

\S4method{chronologies}{sites}(x)
}
\arguments{
\item{x}{sites object that contains chronologies}
}
\value{
object of class \code{chronologies}

chronologies object with all chronologies used.
}
\description{
The class for chronologies from the
Neotoma Paleoecology Database. A single collection unit may
have one or more chronology. The individual chronology
classes are grouped into an S4 \code{chronologies} class.
}
