% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{chronologies_classes}
\alias{chronologies_classes}
\alias{chronology-class}
\alias{chronologies-class}
\title{S4 class for \code{chronologies} information}
\value{
object of class \code{chronologies}
}
\description{
The class for chronologies from the
Neotoma Paleoecology Database. A single collection unit may
have one or more chronology. The individual chronology
classes are grouped into an S4 \code{chronologies} class.
}
