% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{chronology-class}
\alias{chronology-class}
\alias{chronology}
\title{S4 class for chronologies information}
\description{
The grouped class for chronologies
frrom the Neotoma Paleoecology Database.
}
