% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{cite_data,sites-method}
\alias{cite_data,sites-method}
\title{Obtain data citations from records.}
\usage{
\S4method{cite_data}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\description{
Obtain data citations from records.
}
