% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/site-methods.R
\name{cite_data}
\alias{cite_data}
\alias{cite_data,site-method}
\title{Generate a data citation from a Neotoma2 object.}
\usage{
cite_data(x)

\S4method{cite_data}{site}(x)
}
\arguments{
\item{x}{Object with DOIs associated to it.}
}
\value{
data.frame with citation data
}
\description{
The function, applied to a data object with a valid dataset,
will return a properly formatted data citation for the record.
}
