% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{collunits-class}
\alias{collunits-class}
\alias{collunits}
\title{An S4 class for site information from the Neotoma Paleoecology Database.}
\description{
An S4 class for site information from the Neotoma Paleoecology Database.
}
