% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{collunits,site-method}
\alias{collunits,site-method}
\title{Extract \code{collunits} from a \code{site} object.}
\usage{
\S4method{collunits}{site}(object)
}
\arguments{
\item{object}{A \code{site} object}
}
\value{
\code{collunits} from a \code{site} object
}
\description{
Extract \code{collunits} from a \code{site} object.
}
