% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{collunits,sites-method}
\alias{collunits,sites-method}
\title{Extract datasets from a sites object.}
\usage{
\S4method{collunits}{sites}(object)
}
\arguments{
\item{object}{A sites object}
}
\description{
Extract datasets from a sites object.
}
