% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{collunits_classes}
\alias{collunits_classes}
\alias{collunit-class}
\alias{collunits-class}
\title{S4 class for \verb{collection units} information.}
\value{
object of class \code{collunits}
}
\description{
A \verb{collection unit} represents a collection event from within
a \code{site}. For example, a lake sediment core, or a single dig site within an
archaeological site.
}
