% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{contacts_classes}
\alias{contacts_classes}
\alias{contact-class}
\alias{contacts-class}
\title{An S4 class for multi-contact information from the Neotoma
Paleoecology Database.}
\value{
object of class \code{contact}
}
\description{
An unordered list of individual S4 \code{contact} objects.
}
\examples{
new("contact", familyname = "Goring", givennames = "Simon J.")
}
