% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{coordinates,sites-method}
\alias{coordinates,sites-method}
\title{Return the latitude and logitude of sites}
\usage{
\S4method{coordinates}{sites}(obj, ...)
}
\arguments{
\item{x}{sites object 1}
}
\description{
Return the latitude and logitude of sites
}
