% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{coordinates}
\alias{coordinates}
\title{Obtain coordinates from a sites object.}
\usage{
coordinates(obj, ...)
}
\arguments{
\item{object}{A sites object}
}
\description{
Obtain coordinates from a sites object.
}
