% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/site-methods.R
\name{count}
\alias{count}
\alias{count,sites-method}
\alias{count,site-method}
\title{Count the number of sites, datasets, or collectionunits associated with a site object.}
\usage{
count(x, level = "sites")

\S4method{count}{sites}(x, level = c("sites", "collunits", "datasets"))

\S4method{count}{site}(x, level = c("sites", "collunits", "datasets"))
}
\arguments{
\item{x}{A \code{sites} or \code{site} object.}

\item{level}{The level at which you want to count. Can be sites, collunits, or datasets. Default is sites.}
}
\value{
A count of the specified level in the object
}
\description{
Count the number of sites, datasets, or collectionunits associated with a site object.
}
