% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\title{An S4 class for site information from the Neotoma Paleoecology Database.}
\description{
An S4 class for site information from the Neotoma Paleoecology Database.
}
