% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\docType{class}
\name{datasets-class}
\alias{datasets-class}
\alias{datasets}
\title{S4 class for datasets information}
\description{
The grouped class for datasets from
the Neotoma Paleoecology Database.
}
