% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{datasets,collunit-method}
\alias{datasets,collunit-method}
\title{Extract datasets from a collunit object.}
\usage{
\S4method{datasets}{collunit}(object)
}
\arguments{
\item{object}{A collunits object}
}
\description{
Extract datasets from a collunit object.
}
