% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_subset-methods.R, R/subset-methods.R
\name{datasets,collunits-method}
\alias{datasets,collunits-method}
\title{Extract datasets from a collunits object.}
\usage{
\S4method{datasets}{collunits}(object)

\S4method{datasets}{collunits}(object)
}
\arguments{
\item{object}{A collunits object}
}
\description{
Extract datasets from a collunits object.

Extract datasets from a collunits object.
}
