% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{datasets,sites-method}
\alias{datasets,sites-method}
\title{Extract \code{datasets} from a \code{sites} object.}
\usage{
\S4method{datasets}{sites}(object)
}
\arguments{
\item{object}{A \code{sites} object}
}
\value{
\code{datasets} from a \code{sites} object
}
\description{
Extract \code{datasets} from a \code{sites} object.
}
