% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R,
%   R/02_genericDefinitions.R, R/subset-methods.R
\docType{class}
\name{datasets}
\alias{datasets}
\alias{datasets-class}
\alias{datasets,collunit-method}
\alias{datasets,sites-method}
\alias{datasets,site-method}
\title{S4 class for \code{datasets} information}
\usage{
datasets(object)

\S4method{datasets}{collunit}(object)

\S4method{datasets}{sites}(object)

\S4method{datasets}{site}(object)
}
\arguments{
\item{object}{A sites object}
}
\value{
object of class \code{datasets}

datasets object specific to the metadata contained in datasets
}
\description{
The standard object class for \code{datasets}
from the Neotoma Paleoecology Database.

If the sites object contains datasets, then the datasets
will be returned. If the sites object does not contain datasets then
the user can apply \code{get_datasets()} to the object.
}
