% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{datasets-class}
\title{S4 class for \code{datasets} information}
\value{
object of class \code{datasets}
}
\description{
The standard object class for \code{datasets}
from the Neotoma Paleoecology Database.
}
