% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{doi,publication-method}
\alias{doi,publication-method}
\title{Get a publication DOI.}
\usage{
\S4method{doi}{publication}(object)
}
\arguments{
\item{x}{A \code{publication} object.}
}
\description{
Get a publication DOI.
}
