% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{doi,site-method}
\alias{doi,site-method}
\title{Obtain dataset DOIs from records.}
\usage{
\S4method{doi}{site}(x)
}
\arguments{
\item{x}{site object}
}
\description{
Obtain dataset DOIs from records.
}
