% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{doi,sites-method}
\alias{doi,sites-method}
\title{Obtain dataset DOIs from records.}
\usage{
\S4method{doi}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\description{
Obtain dataset DOIs from records.
}
