% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter,sites-method}
\alias{filter,sites-method}
\title{filter}
\usage{
\S4method{filter}{sites}(x, ...)
}
\arguments{
\item{x}{A site, dataset or download.}

\item{...}{arguments to filter by: lat, long,
elev, datasettype}
}
\description{
filter
}
