% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_nulls.r
\name{fix_null}
\alias{fix_null}
\title{Fix null values from API responses}
\usage{
fix_null(x)
}
\description{
API responses from the Neotoma API return many fields
as NULL values. This function turns the NULL values into logical NAs.
}
