% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest.R
\name{get_closest}
\alias{get_closest}
\title{Get closest}
\usage{
get_closest(x, n, buffer, ...)
}
\arguments{
\item{x}{A vector long/lat pair, or a dataset, site or download.}

\item{n}{The maximum number of records to return (in the case of ties the return may be larger)}

\item{buffer}{The size of the buffer for dataset search (in kilometers)}

\item{...}{optional arguments to pass into \code{get_dataset}.}
}
\description{
Get closest
}
