% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts.default}
\alias{get_contacts.default}
\title{Get contact information for Neotoma contributors}
\usage{
\method{get_contacts}{default}(x, ...)
}
\arguments{
\item{x}{integer A contact ID}

\item{...}{Other options to pass to \code{get_contacts()}.}
}
\description{
Get contact information for Neotoma contributors
}
