% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts.R
\name{get_contacts.numeric}
\alias{get_contacts.numeric}
\title{Get contact information for Neotoma contributors}
\usage{
\method{get_contacts}{numeric}(x, ...)
}
\description{
Get contact information for Neotoma contributors
}
