% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.default}
\alias{get_datasets.default}
\title{Get Dataset Default}
\usage{
\method{get_datasets}{default}(x, ..., complete_data = FALSE)
}
\arguments{
\item{x}{Use a single number to extract site information}

\item{...}{contactid, datasettype,
altmin, altmax, loc, ageyoung, ageold, ageof}
}
\description{
Get Dataset Default
}
