% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.sites}
\alias{get_datasets.sites}
\title{Get Dataset from a \code{sites} object.}
\usage{
\method{get_datasets}{sites}(sites, verbose = 0)
}
\arguments{
\item{sites}{An object of class \code{sites}.}

\item{verbose}{Should the function post text as it executes?}
}
\description{
Get Dataset from a \code{sites} object.
}
