% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads.numeric}
\alias{get_downloads.numeric}
\title{get_downloads}
\usage{
\method{get_downloads}{numeric}(x, ...)
}
\arguments{
\item{x}{Use a single number to extract site information}

\item{...}{arguments in ellipse form}

\item{verbose}{Should text be printed during the download process?}
}
\value{
The function returns either a single item of class
\code{"try-error"} describing the reason for failure
(either misdefined parameters or an error from the Neotoma API),
or a table of sites, with rows corresponding to the number of
individual sites and datasets returned by the Neotoma API.
}
\description{
get_downloads
}
