% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual.R
\name{get_manual}
\alias{get_manual}
\title{get_manual}
\usage{
get_manual()
}
\value{
NULL side effect for opening browser with the manual
}
\description{
Open up the Neotoma manual homepage.
}
\examples{
{
# This call does not work from `source()` calls or in testing.
# interactive() just lets us know you are interacting with the console:
if (interactive()) {
 get_manual()
}
}

}
\author{
Simon Goring \email{goring@wisc.edu}
}
