% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications.default}
\alias{get_publications.default}
\title{Get contact information for Neotoma contributors}
\usage{
\method{get_publications}{default}(...)
}
\description{
Get contact information for Neotoma contributors
}
