% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site.R
\name{get_site.default}
\alias{get_site.default}
\title{Get Site Information for Fossil Sites}
\usage{
\method{get_site}{default}(sitename = NA, lat = NA, ...)
}
\arguments{
\item{sitename}{gets sitename}
}
\description{
Get Site Information for Fossil Sites
}
