% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site.R
\name{get_site}
\alias{get_site}
\title{Get Site Information for Fossil Sites}
\usage{
get_site(x = NA, ...)
}
\arguments{
\item{x}{integer A contact ID}

\item{contactname}{A full or partial name for an individual contributor to the database.}

\item{familyname}{The full or partial last name for an individual contributor to the database.}
}
\description{
Get Site Information for Fossil Sites
}
