% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.datasets}
\alias{get_sites.datasets}
\title{Get Site Datasets}
\usage{
\method{get_sites}{datasets}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dataset}.}

\item{...}{arguments in ellipse form}
}
\description{
Get Site Datasets
}
\examples{
get_sites.datasets(24)
}
