% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.numeric}
\alias{get_sites.numeric}
\title{Get Site Numeric}
\usage{
\method{get_sites}{numeric}(siteid, ...)
}
\arguments{
\item{x}{Use a single number to extract site information}
}
\description{
Get Site Numeric
}
\examples{
get_sites(24)

A site object containing 1 sites and 6 parameters. 
| siteid  | sitename       |    lat    |   long   | elev |
| :------ |:--------------:|:---------:|:--------:|-----:|
| 24      | Alexander Lake | -60.58333 | 53.33333 |  143 |
}
