% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.sites}
\alias{get_sites.sites}
\title{Get Site Information for Fossil Sites from a Set of Sites}
\usage{
\method{get_sites}{sites}(x, ...)
}
\arguments{
\item{x}{The numeric site ID from Neotoma}

\item{...}{accepted arguments if numeric all_data}
}
\description{
Get Site Information for Fossil Sites from a Set of Sites
}
\examples{
\dontrun{
## Find all sites using a set of prior sites:
char_sites <- get_sites(taxa = "charcoal")
<<<<<<< Updated upstream
sites <- get_sites(char_sites)
# Note that each of these have a length of 25 (sites)
# because we didn't use a `limit`, however, we can see
# that `sites` has a full compliment of datasets:
length(datasets(sites)) > length(datasets(char_sites))
pollen_coloc <- sites \%>\% filter(datasettype == 'pollen')
char_coloc <- char_sites \%>\% filter(siteid \%in\% getids(pollen_coloc)$siteid)
pol_char <- c(pollen_coloc, char_coloc)
=======
pollen_coloc <- get_sites(char_sites, datasettype = "pollen")
char_coloc <- char_sites \%>\% filter(siteid \%in\% getids(pollen_coloc)$siteid)
pol_char <- c(pollen_coloc, char_coloc) \%>\% clean()
>>>>>>> Stashed changes
}
}
