% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens.default}
\alias{get_specimens.default}
\title{Get Specimen datasetid}
\usage{
\method{get_specimens}{default}(...)
}
\arguments{
\item{...}{Pass argument datasetid and the corresponding datasetid}
}
\description{
Get Specimen datasetid
}
