% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens.numeric}
\alias{get_specimens.numeric}
\title{Get Specimen Numeric}
\usage{
\method{get_specimens}{numeric}(x)
}
\arguments{
\item{x}{Use a single number to extract site information}
}
\description{
Get Specimen Numeric
}
