% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens}
\alias{get_specimens}
\title{get_specimens}
\usage{
get_specimens(x = NA, ...)
}
\arguments{
\item{x}{Use a single specimenid}

\item{...}{Additional terms passed to get_specimens, most common datasetid}
}
\value{
The function returns a specimens list
}
\description{
Information for Specimens
}
\examples{
\donttest{
# To find specimen with ID 7:
my_specimens <- get_specimens(7)
# To find specimens in datasetid 41610
my_specimens2 <- get_specimens(datasetid = 41610)
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
