% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens.sites}
\alias{get_specimens.sites}
\title{Get Specimen Sites}
\usage{
\method{get_specimens}{sites}(x)
}
\arguments{
\item{x}{Use a single number to extract site information}
}
\description{
Get Specimen Sites
}
