% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speleothems.R
\name{get_speleothems.numeric}
\alias{get_speleothems.numeric}
\title{Get Speleothem Data from a Sites object}
\usage{
\method{get_speleothems}{numeric}(x, ...)
}
\arguments{
\item{x}{The numeric dataset ID from Neotoma}

\item{...}{accepted arguments if numeric all_data}
}
\value{
The function returns either a single item of class \code{"try-error"}
describing the reason for failure (either misdefined parameters or an error
from the Neotoma API), or a sies object with speleothem data.
}
\description{
Get Speleothem Data from a Sites object
}
\examples{
{
## Find speleothems by numeric datasetid:
sites <- get_seleothem(seq(1,3))
}
}
