% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R, R/get_taxon.R
\name{get_taxon.numeric}
\alias{get_taxon.numeric}
\title{Get Taxa Numeric}
\usage{
\method{get_taxon}{numeric}(x, ...)

\method{get_taxon}{numeric}(x, ...)
}
\arguments{
\item{x}{Use a taxon ID to extract sites information}

\item{...}{Additional parameters to get_taxa}
}
\value{
\code{sites} object with requested \code{taxa}

\code{sites} object with requested \code{taxa}
}
\description{
Get Taxa Numeric

Get Taxa Numeric
}
\examples{
\donttest{
allds <- get_datasets(1:3)
}
\donttest{
allds <- get_datasets(1:3)
}
}
