% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.collunit}
\alias{getids.collunit}
\title{Get object IDs from a single collectionunit.}
\usage{
\method{getids}{collunit}(x, order = TRUE)
}
\arguments{
\item{x}{A neotoma2 \code{collunit} object.}

\item{order}{sort items by siteid, collunitid, datasetid}
}
\description{
From a collecitonunit object, return the
collectionunit and dataset ids.
}
\examples{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
collunitids <- getids(collunits(marion)[[1]])
}
