% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.collunits}
\alias{getids.collunits}
\title{Get object IDs from collectionunits.}
\usage{
\method{getids}{collunits}(x, order = TRUE)
}
\arguments{
\item{x}{A Neotoma2 \code{collunits} object.}

\item{order}{sort items by siteid, collunitid, datasetid}
}
\description{
From a set of collecitonunitobjects, return the
collectionunit and dataset ids.
}
\examples{
\dontrun{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
collunitids <- getids(collunits(marion))
}
}
