% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.site}
\alias{getids.site}
\title{Get object IDs from sites}
\usage{
\method{getids}{site}(x, order = TRUE)
}
\arguments{
\item{x}{A neotoma2 \code{sites} object.}
}
\description{
Get object IDs from sites
}
