% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.sites}
\alias{getids.sites}
\title{Get object IDs from sites}
\usage{
\method{getids}{sites}(x, order = TRUE)
}
\arguments{
\item{x}{A Neotoma2 \code{sites} object.}

\item{order}{sort items by siteid, collunitid, datasetid}
}
\value{
\code{data.frame} containing \code{siteid}, \code{datasetid}, and \code{collunitid}
}
\description{
Get object IDs from sites
}
