% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{hash,chronology-method}
\alias{hash,chronology-method}
\title{Hash a chronology object}
\usage{
\S4method{hash}{chronology}(x)
}
\arguments{
\item{object}{chronology object}
}
\description{
Hash a Neotoma chronology object
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunit[[1]]$chronology[[1]])
}
