% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{hash,collunit-method}
\alias{hash,collunit-method}
\title{Hash a collunit object}
\usage{
\S4method{hash}{collunit}(x)
}
\arguments{
\item{object}{collunit object}
}
\description{
Hash a Neotoma collection unit object to add as an attribute
returning the hashed siteid, sitename, latitude, longitude and altitude.
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunit[[1]])
}
