% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{hash,dataset-method}
\alias{hash,dataset-method}
\title{Hash a dataset object}
\usage{
\S4method{hash}{dataset}(x)
}
\arguments{
\item{object}{dataset object}
}
\description{
Hash a Neotoma dataset object to add as an attribute
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunits[[1]]$datasets[[1]])
}
