% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{hash,sample-method}
\alias{hash,sample-method}
\title{Hash a sample object}
\usage{
\S4method{hash}{sample}(x)
}
\arguments{
\item{object}{sample object}
}
\description{
Hash a Neotoma sample object
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunit[[1]]$dataset[[1]]$samples[[1]])
}
