% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{hash,site-method}
\alias{hash,site-method}
\title{Hash a site object}
\usage{
\S4method{hash}{site}(x)
}
\arguments{
\item{object}{site object}
}
\description{
Hash a Neotoma site object to add as an attribute
returning the hashed siteid, sitename, latitude, longitude and altitude.
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]])
}
