% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{length,chronologies-method}
\alias{length,chronologies-method}
\title{Length Method chronologies}
\usage{
\S4method{length}{chronologies}(x)
}
\arguments{
\item{x}{chronologies object}
}
\value{
integer describing length
}
\description{
Length Method chronologies
}
