% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{length,collunits-method}
\alias{length,collunits-method}
\title{Length Method collunits}
\usage{
\S4method{length}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\value{
length of a \code{collunits} object
}
\description{
Length Method collunits
}
