% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{length,datasets-method}
\alias{length,datasets-method}
\title{Length Method Sites}
\usage{
\S4method{length}{datasets}(x)
}
\arguments{
\item{x}{datasets object}
}
\description{
Length Method Sites
}
