% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{length,samples-method}
\alias{length,samples-method}
\title{Length Method samples}
\usage{
\S4method{length}{samples}(x)
}
\arguments{
\item{x}{samples object}
}
\value{
\code{int} representing the length of \verb{samples object}
}
\description{
Length Method samples
}
