% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{length,sites-method}
\alias{length,sites-method}
\title{Length Method Sites}
\usage{
\S4method{length}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
\code{int} with the length of sites object
}
\description{
Length Method Sites
}
