% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{length,specimens-method}
\alias{length,specimens-method}
\title{Length Method specimens}
\usage{
\S4method{length}{specimens}(x)
}
\arguments{
\item{x}{specimens object}
}
\value{
\code{int} with length of \code{specimens} object
}
\description{
Length Method specimens
}
