% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{length,speleothems-method}
\alias{length,speleothems-method}
\title{Length Method speleothems}
\usage{
\S4method{length}{speleothems}(x)
}
\arguments{
\item{x}{speleothems object}
}
\value{
\code{int} that showcases the length of a \code{speleothems} object
}
\description{
Length Method speleothems
}
