% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{length,taxa-method}
\alias{length,taxa-method}
\title{Length Method taxa}
\usage{
\S4method{length}{taxa}(x)
}
\arguments{
\item{x}{taxa object}
}
\value{
\code{int} that showcases the length of a \code{taxa} object
}
\description{
Length Method taxa
}
