% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/dataset-methods.R, R/publication-methods.R, R/samples-methods.R,
%   R/site-methods.R, R/speleothem-methods.R, R/taxon-methods.R
\name{length,chronologies-method}
\alias{length,chronologies-method}
\alias{length,collunits-method}
\alias{length,datasets-method}
\alias{length,publications-method}
\alias{length,samples-method}
\alias{length}
\alias{length,sites-method}
\alias{length,speleothems-method}
\alias{length,taxa-method}
\title{length}
\usage{
\S4method{length}{chronologies}(x)

\S4method{length}{collunits}(x)

\S4method{length}{datasets}(x)

\S4method{length}{publications}(x)

\S4method{length}{samples}(x)

\S4method{length}{sites}(x)

\S4method{length}{speleothems}(x)

\S4method{length}{taxa}(x)
}
\arguments{
\item{x}{\code{neotoma2} object}
}
\value{
\code{int} representing length of a \code{neotoma2} object
}
\description{
length
}
