% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/neotoma-methods.R
\docType{class}
\name{missingOrNULL-class}
\alias{missingOrNULL-class}
\title{c Method - Combine objects, including NULL}
\description{
c Method - Combine objects, including NULL

c Method - Combine objects, including NULL
}
