% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{names,collunit-method}
\alias{names,collunit-method}
\title{Get slot names}
\usage{
\S4method{names}{collunit}(x)
}
\arguments{
\item{x}{A collection unit object.}
}
\value{
NULL. Shows the names of the slots
}
\description{
Get all names for named elements within a \code{collunit} object.
}
