% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{names,contact-method}
\alias{names,contact-method}
\title{Get names of contacts slots}
\usage{
\S4method{names}{contact}(x)
}
\arguments{
\item{x}{A contact object.}
}
\value{
names of slots
}
\description{
Get names of contacts slots
}
