% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{names,dataset-method}
\alias{names,dataset-method}
\title{Get slot names}
\usage{
\S4method{names}{dataset}(x)
}
\arguments{
\item{x}{A dataset object.}
}
\value{
\code{list} with all names of \code{dataset} slots
}
\description{
Get all names for named elements within a \code{dataset} object.
}
