% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{names,site-method}
\alias{names,site-method}
\title{Get slot names}
\usage{
\S4method{names}{site}(x)
}
\arguments{
\item{x}{A site object.}
}
\value{
names of the slots of a \code{site} object
}
\description{
Get all names for named elements within a \code{site} object.
}
