% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{names,specimen-method}
\alias{names,specimen-method}
\title{Get slot names}
\usage{
\S4method{names}{specimen}(x)
}
\arguments{
\item{x}{A specimen object.}
}
\value{
names of the slots of a \code{site} object
}
\description{
Get all names for named elements within a \code{specimen} object.
}
