% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{names,taxon-method}
\alias{names,taxon-method}
\title{Get slot names}
\usage{
\S4method{names}{taxon}(x)
}
\arguments{
\item{x}{A taxon object.}
}
\value{
\code{list} with all names of \code{taxon} slots
}
\description{
Get all names for named elements within a \code{taxon} object.
}
