% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/contact-methods.R,
%   R/dataset-methods.R, R/publication-methods.R, R/site-methods.R,
%   R/speleothem-methods.R, R/taxon-methods.R
\name{names,collunit-method}
\alias{names,collunit-method}
\alias{names,contact-method}
\alias{names,dataset-method}
\alias{names,publication-method}
\alias{names,publications-method}
\alias{names}
\alias{names,sites-method}
\alias{names,speleothem-method}
\alias{names,taxon-method}
\title{Get a neotoma2 object's slot names}
\usage{
\S4method{names}{collunit}(x)

\S4method{names}{contact}(x)

\S4method{names}{dataset}(x)

\S4method{names}{publication}(x)

\S4method{names}{publications}(x)

\S4method{names}{site}(x)

\S4method{names}{sites}(x)

\S4method{names}{speleothem}(x)

\S4method{names}{taxon}(x)
}
\arguments{
\item{x}{A neotoma2 object.}
}
\value{
NULL.
}
\description{
Get all names for elements' slots within a \code{collunit} object.
}
