% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseURL.R
\name{newURL}
\alias{newURL}
\title{Format API call to Neotoma from call arguments}
\usage{
newURL(baseurl, args, ...)
}
\arguments{
\item{baseurl}{The base URL for the Neotoma API}

\item{args}{The set of query arguments to be passed to the API}

\item{...}{Any additional arguments to be passed to the function.}
}
\value{
A properly formatted URL.
}
\description{
Take a set of arguments from the Neotoma2 package and produce
the appropriate URL to the Neotoma v2.0 API.
This is an internal function used by \code{parseURL()}.
}
