% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{pager}
\alias{pager}
\title{pager}
\usage{
pager(response, response_url, ...)
}
\arguments{
\item{response}{The result of the API call}

\item{response_url}{the url field of the http call}

\item{...}{If TRUE, modify the response to contain all datapoints}
}
\description{
An internal helper function to count through all offsets/limits
}
