% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{parse_location}
\alias{parse_location}
\title{parse_location}
\usage{
parse_location(x)
}
\arguments{
\item{x}{location object}
}
\description{
Retrieve location from datasdets WT, geojson, bounding box
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
