% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_body.R
\name{parsebody}
\alias{parsebody}
\title{parse_body}
\usage{
parsebody(x, ...)
}
\arguments{
\item{x}{The HTTP path for the particular API call.}

\item{...}{Any query parameters passed from the function calling}
}
\value{
\code{JSON} object to parse as a body in a HTTP request
}
\description{
An internal helper function to parse the body of POST API requests
}
\author{
Socorro Dominguez
}
