% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseURL.R
\name{parseURL}
\alias{parseURL}
\title{parseURL}
\usage{
parseURL(x, use = "neotoma", all_data = FALSE, fetch_fn = NULL, ...)
}
\arguments{
\item{x}{The HTTP/S path for the particular API call.}

\item{use}{Uses the Neotoma server by default ("neotoma"), but supports either the
development API server ("dev") or a local server ("local").}

\item{all_data}{If TRUE return all possible API calls}

\item{...}{Any query parameters passed from the calling function.}
}
\value{
\code{list} with cleaned and parsed data from HTTP request
}
\description{
An internal helper function used to connect to the Neotoma API
in a standard manner, and to provide basic validation of any response.
}
\author{
Socorro Dominguez \email{s.dominguez@ht-data.com}

Simon Goring \email{goring@wisc.edu}
}
