% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pingNeotoma.r
\name{pingNeotoma}
\alias{pingNeotoma}
\title{pingNeotoma}
\usage{
pingNeotoma(server = "neotoma")
}
\arguments{
\item{server}{One of \code{localhost:PORT} (where \code{PORT} is a valid
numeric port), \code{neotoma} or \code{dev}.}
}
\value{
A valid HTTP status code or returns an error if a connection
is refused.
}
\description{
A quick function to test whether or not the Neotoma Database
API is currently running.
}
\examples{
{
test_connection <- pingNeotoma("neotoma")
}
}
