% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{plot,sites-method}
\alias{plot,sites-method}
\title{Plot site coordinates using a basic plot.}
\usage{
\S4method{plot}{sites}(x, y, ...)
}
\arguments{
\item{x}{sites object}

\item{...}{Additional parameters associated with the call.}
}
\value{
\code{plot} object with site coordinates.
}
\description{
Plot site coordinates using a basic plot.
}
