% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{plot}
\alias{plot}
\alias{plot,sites-method}
\alias{plot,sites,ANY-method}
\alias{plot,site,ANY-method}
\alias{plot,site-method}
\title{Plot site coordinates using a basic plot.}
\usage{
\S4method{plot}{sites,ANY}(x, y, ...)

\S4method{plot}{site,ANY}(x, y, ...)
}
\arguments{
\item{x}{sites object}

\item{y}{\emph{Ignored.}}

\item{...}{Additional parameters associated with the call.}
}
\value{
\code{plot} object with site coordinates.
}
\description{
Plot site coordinates using a basic plot.
}
