% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeaflet.R
\name{plotLeaflet,site-method}
\alias{plotLeaflet,site-method}
\title{plotLeaflet}
\usage{
\S4method{plotLeaflet}{site}(object, save_im = FALSE, path = "")
}
\arguments{
\item{object}{Site object to plot}

\item{save_im}{save output}

\item{path}{location where output should be saved in. save_im must be TRUE}
}
\description{
Plot a site on a leaflet map
}
