% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeaflet.R
\name{plotLeaflet,sites-method}
\alias{plotLeaflet,sites-method}
\title{plotLeaflet}
\usage{
\S4method{plotLeaflet}{sites}(object)
}
\arguments{
\item{object}{Sites object to plot}
}
\value{
\code{leaflet} map
}
\description{
Plot sites on a leaflet map
}
\examples{
\donttest{
# Note that by default the limit for queries is 25 records:
modernSites <- get_sites(keyword = "Modern")
plotLeaflet(modernSites)
}
}
