% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubAuthors.R
\name{pubAuthors}
\alias{pubAuthors}
\title{Reformat API author list to Neotoma contacts.}
\usage{
pubAuthors(x)
}
\arguments{
\item{x}{A list coming from the \code{author} element of the Neotoma publicaiton API}
}
\description{
Reformat API author list to Neotoma contacts.
}
