% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{publication-class}
\alias{publication-class}
\title{An S4 class for a single Neotoma publication.}
\description{
A publication is liked to an individual Neotoma dataset object
They are grouped using an S4 \code{publications} class.
}
