% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{repositories-class}
\alias{repositories-class}
\title{S4 class for repositories information}
\value{
object of class \code{repositories}
}
\description{
The grouped class for repositories from
the Neotoma Paleoecology Database.
}
