% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{repositories}
\alias{repositories}
\alias{repositories-class}
\title{S4 class for \code{repository} information}
\value{
object of class \code{repository}
}
\description{
The standard object class for \code{repository}
from the Neotoma Paleoecology Database.
}
