% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{repository-class}
\alias{repository-class}
\title{S4 class for repository information}
\description{
The standard object class for repository
from the Neotoma Paleoecology Database.
}
