% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asamples-methods.R
\docType{class}
\name{samples-class}
\alias{samples-class}
\alias{samples}
\title{S4 class for datasets information}
\description{
The grouped class for datasets from
the Neotoma Paleoecology Database.
}
