% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,collunit-method}
\alias{samples,collunit-method}
\title{samples}
\usage{
\S4method{samples}{collunit}(x)
}
\arguments{
\item{x}{collunit object}
}
\description{
Obtain elements from collunit
}
\examples{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
pollen <- samples(collunits(marion)[[1]])
plot(value ~ I(-1 * age),
     data = pollen[pollen$variablename == "Cupressaceae",],
xlab = "Years before present",
ylab = "Cupressaceae pollen count")
}
