% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,collunits-method}
\alias{samples,collunits-method}
\title{samples}
\usage{
\S4method{samples}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\description{
Obtain elements from collunits
}
