% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,site-method}
\alias{samples,site-method}
\title{samples}
\usage{
\S4method{samples}{site}(x)
}
\arguments{
\item{x}{site object}
}
\description{
Obtain elements on the samples level
}
\examples{
\dontrun{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
pollen <- samples(marion)
plot(value ~ I(-1 * age), 
     data = pollen[pollen$variablename == "Cupressaceae",],
xlab = "Years before present",
ylab = "Cupressaceae pollen count")
}
}
