% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,sites-method}
\alias{samples,sites-method}
\title{samples}
\usage{
\S4method{samples}{sites}(x, chronname = NULL)
}
\arguments{
\item{x}{sites object}

\item{chronname}{Optional character name of a chronology to report ages
against. When \code{NULL} (default) the Neotoma default-chronology precedence
is applied. If the named chronology is not present for a dataset, the
default is used and a warning is issued.}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain all samples within a sites object
}
\examples{
\dontrun{
# Get full data download from API and create a long table with samples data.
dw <- get_downloads(1)
pollen <- samples(dw)
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}

Simon Goring \email{goring@wisc.edu}
}
