% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R,
%   R/02_genericDefinitions.R, R/samples.R
\docType{class}
\name{samples-class}
\alias{samples-class}
\alias{samples}
\alias{samples,site-method}
\alias{samples,collunits-method}
\alias{samples,collunit-method}
\title{Obtain samples from a record or multiple records.}
\usage{
samples(x)

\S4method{samples}{site}(x)

\S4method{samples}{collunits}(x)

\S4method{samples}{collunit}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at sample level
}
\description{
Obtain samples from a record or multiple records.
}
