% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{selectMatch,publication,numeric-method}
\alias{selectMatch,publication,numeric-method}
\title{Select the best match (between a local record and a Neotoma match)}
\usage{
\S4method{selectMatch}{publication,numeric}(x, n)
}
\arguments{
\item{x}{A \code{publication} object}

\item{n}{The match number.}
}
\value{
the best match to the selected publication.
}
\description{
Select the best match (between a local record and a Neotoma match)
}
