% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R,
%   R/publication-methods.R
\name{selectMatch}
\alias{selectMatch}
\alias{selectMatch,publication,logical-method}
\title{Select the best match for an object.}
\usage{
selectMatch(x, n)

\S4method{selectMatch}{publication,logical}(x, n)
}
\arguments{
\item{x}{object}

\item{n}{n elements that are a best match}
}
\value{
attr Select the match between a local record and a Neotoma match
}
\description{
Select the best match for an object.
}
