% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_chron.R, R/set_chronology.R
\name{set_chronology}
\alias{set_chronology}
\title{set Site Information for Fossil Sites}
\usage{
set_chronology(
  x = NA,
  notes = NA_character_,
  contact = list(),
  agemodel = NA_character_,
  ageboundolder = NA_integer_,
  ageboundyounger = NA_integer_,
  isdefault = NA_integer_,
  dateprepared = as.Date(character(0)),
  modelagetype = NA_character_,
  chronologyname = NA_character_,
  chroncontrols = data.frame()
)

set_chronology(
  x = NA,
  notes = NA_character_,
  contact = list(),
  agemodel = NA_character_,
  ageboundolder = NA_integer_,
  ageboundyounger = NA_integer_,
  isdefault = NA_integer_,
  dateprepared = as.Date(character(0)),
  modelagetype = NA_character_,
  chronologyname = NA_character_,
  chroncontrols = data.frame()
)
}
\arguments{
\item{x}{Object to be set as a chronology}

\item{notes}{<- notes}

\item{contact}{<- contact}

\item{agemodel}{<- agemodel}

\item{ageboundolder}{<- ageboundolder}

\item{ageboundyounger}{<- ageboundyounger}

\item{isdefault}{<- isdefault}

\item{dateprepared}{<- dateprepared}

\item{modelagetype}{<- modelagetype}

\item{chronologyname}{<- chronologyname}

\item{chroncontrols}{<- chroncontrols}
}
\description{
set Site Information for Fossil Sites

set chronology information for a new record.
}
\examples{
\dontrun{
# Create a site called "My Lake", to
my_chron <- set_site(sitename = "My Lake",
                    geography = x,
                    description = "my lake",
                    altitude = 30)
}

\dontrun{
# Create a site called "My Lake", to
x = st_as_sf(st_sfc(st_point(c(5,5))))
my_site <- set_site(sitename = "My Lake",
                    geography = x,
                    description = "my lake",
                    altitude = 30)
}
}
