% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{set_default,chronologies-method}
\alias{set_default,chronologies-method}
\title{Change the default age model for a record.}
\usage{
\S4method{set_default}{chronologies}(x, n)
}
\arguments{
\item{x}{A chronologies object.}

\item{n}{The particular chronology to be used as the default.}
}
\value{
\code{chronologies} object with a new defaulted \code{chronology}
}
\description{
Change the default age model for a record.
}
