% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_publications.R
\name{set_publication}
\alias{set_publication}
\title{Create a new publication (or publication set)}
\usage{
set_publication(...)
}
\arguments{
\item{publicationtype}{A text string identifying the publication type within the Neotoma database.}

\item{year}{The year of publication.}

\item{citation}{A full text citation for the article.}

\item{articletitle}{The title of the article.}

\item{journal}{The journal in which the article was published.}

\item{volume}{The journal volume.}

\item{issue}{The journal issue.}

\item{pages}{The pages of the journal.}

\item{citationnumber}{How many times has the paper been cited?}

\item{doi}{A DOI for the record.}

\item{booktitle}{The title of the book (if the publication is a book)}

\item{numvolumes}{The number of book volumes (if a series)}

\item{edition}{The book edition.}

\item{volumetitle}{The title of the volume (in a published series)}

\item{seriestitle}{The title of the series.}

\item{seriesvolume}{The series volume.}

\item{publisher}{The publisher.}

\item{url}{Publication URL}

\item{city}{City of publication.}

\item{state}{State of publication.}

\item{country}{Country of publication.}

\item{originallanguage}{Original language of publication.}

\item{notes}{Publication notes.}
}
\description{
A function to create new publication objects by hand.
}
