% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sample.R
\name{set_sample}
\alias{set_sample}
\title{set Sample Information}
\usage{
set_sample(
  x = NA,
  ages = df_age,
  igsn = NA_character_,
  datum = df,
  depth = NA_integer_,
  sampleid = NA_integer_,
  thickness = NA_integer_,
  samplename = NA_character_,
  sampleanalyst = analyst_list_helper,
  analysisunitid = NA_integer_,
  analysisunitname = NA_character_
)
}
\arguments{
\item{x}{Object to be set as a sample}

\item{ages}{ages}

\item{igsn}{IGSN character}

\item{datum}{dataframe of datum}

\item{depth}{integer representing depth}

\item{sampleid}{ID for sample}

\item{thickness}{thickness of core}

\item{samplename}{sample's name}

\item{sampleanalyst}{Analyst's contact name}

\item{analysisunitid}{Which analysis unit it is}

\item{analysisunitnamesiteid}{Analysis Unit's name}
}
\description{
set Sample Information
}
\examples{
\dontrun{
# Create a site called "My Lake", to
x = st_as_sf(st_sfc(st_point(c(5,5))))
my_site <- set_site(sitename = "My Lake",
                    geography = x,
                    description = "my lake",
                    altitude = 30)
}
}
