% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_site.R
\name{set_site.default}
\alias{set_site.default}
\title{set Site Information for Fossil Sites}
\usage{
\method{set_site}{default}(
  siteid,
  sitename = NA_character_,
  location = st_sf(st_sfc()),
  description = NA_character_,
  notes = NA_character_,
  collunits = new("collunits")
)
}
\arguments{
\item{siteid}{helps}

\item{sitename}{helps}
}
\description{
set Site Information for Fossil Sites
}
