% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_specimen.R
\name{set_specimen}
\alias{set_specimen}
\title{set Sample Information}
\usage{
set_specimen(
  x = NA,
  datasetid = NA_integer_,
  sampleid = NA_integer_,
  specimenid = NA_integer_,
  repository = new("repository"),
  taxonid = NA_integer_,
  taxonname = NA_character_,
  elementtype = NA_character_,
  symmetry = NA_character_,
  portion = NA_character_,
  sex = NA_character_,
  domesticstatus = NA_character_,
  taphonomictype = NA_character_,
  nisp = NA_integer_,
  preservative = NA_character_,
  maturity = NA_character_,
  samplenotes = NA_character_
)
}
\arguments{
\item{x}{Object to be set as a sample}

\item{sampleid}{ID for sample}

\item{ages}{ages}

\item{igsn}{IGSN character}

\item{datum}{dataframe of datum}

\item{depth}{integer representing depth}

\item{thickness}{thickness of core}

\item{samplename}{sample's name}

\item{sampleanalyst}{Analyst's contact name}

\item{analysisunitid}{Which analysis unit it is}

\item{analysisunitname}{Analysis Unit's name}
}
\description{
set Sample Information
}
\examples{
\dontrun{
# Create a site called "My Lake", to
x = st_as_sf(st_sfc(st_point(c(5,5))))
my_site <- set_site(sitename = "My Lake",
                    geography = x,
                    description = "my lake",
                    altitude = 30)
}
}
