% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNeotoma.R
\name{setNeotoma}
\alias{setNeotoma}
\title{Set Neotoma API Source}
\usage{
setNeotoma(server = "neotoma")
}
\arguments{
\item{server}{One of \code{local} (when the API is running locally on port 3005),
\code{neotoma} or \code{dev}.}
}
\description{
Choose to pull Neotoma data from the main Neotoma server, the development
server or from a local instance of the API.
}
\examples{
# The user is running the API locally using the node/express API
# cloned from github: https://github.com/NeotomaDB/api_nodetest
setNeotoma(server = "neotoma")
}
