% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{show,chronologies-method}
\alias{show,chronologies-method}
\title{Show chronologies object as a dataframe}
\usage{
\S4method{show}{chronologies}(object)
}
\arguments{
\item{object}{chronologies object}
}
\description{
Show chronologies object as a dataframe
}
