% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{show,chronology-method}
\alias{show,chronology-method}
\title{Show Dataset Method}
\usage{
\S4method{show}{chronology}(object)
}
\arguments{
\item{object}{chronology object}
}
\description{
Show Dataset Method
}
