% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{show,collunit-method}
\alias{show,collunit-method}
\title{Show collunit Method}
\usage{
\S4method{show}{collunit}(object)
}
\arguments{
\item{object}{collunit object}
}
\description{
Show collunit Method
}
