% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{show,collunits-method}
\alias{show,collunits-method}
\title{Show Collection Units Method}
\usage{
\S4method{show}{collunits}(object)
}
\arguments{
\item{object}{collunits object}
}
\description{
Show Collection Units Method
}
