% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{show,contact-method}
\alias{show,contact-method}
\title{Show contact object}
\usage{
\S4method{show}{contact}(object)

\S4method{show}{contact}(object)
}
\arguments{
\item{object}{A contact object.}
}
\description{
Show contact object

Show a contact object
}
