% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{show,contacts-method}
\alias{show,contacts-method}
\title{Show a contacts object.}
\usage{
\S4method{show}{contacts}(object)
}
\arguments{
\item{object}{A contacts object.}
}
\value{
null - side effect for printing contacts object
}
\description{
Show a contacts object.
}
