% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{show,dataset-method}
\alias{show,dataset-method}
\title{Show Dataset Method}
\usage{
\S4method{show}{dataset}(object)
}
\arguments{
\item{object}{dataset object}
}
\value{
null - side effect, prints a \code{data.frame} with \code{dataset} metadata
}
\description{
Show Dataset Method
}
