% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{show,publications-method}
\alias{show,publications-method}
\title{Show the contents of a publication object.}
\usage{
\S4method{show}{publications}(object)
}
\arguments{
\item{object}{A \code{publications} object}
}
\value{
NULL - side effect function of printing a data.frame
}
\description{
Show the contents of a publication object.
}
