% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{show,sample-method}
\alias{show,sample-method}
\title{Show the sample information}
\usage{
\S4method{show}{sample}(object)
}
\arguments{
\item{object}{sample object}
}
\description{
Show the sample information
}
