% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{show,site-method}
\alias{show,site-method}
\title{Show Site objects as a dataframe}
\usage{
\S4method{show}{site}(object)
}
\arguments{
\item{object}{site object}
}
\description{
Show Site objects as a dataframe
}
