% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{show,sites-method}
\alias{show,sites-method}
\title{Show result as a brief dataframe - as in Neotoma v1}
\usage{
\S4method{show}{sites}(object)
}
\description{
Show result as a brief dataframe - as in Neotoma v1
}
