% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{show,specimen-method}
\alias{show,specimen-method}
\title{Show Specimen Method}
\usage{
\S4method{show}{specimen}(object)
}
\arguments{
\item{object}{specimen object}
}
\value{
NULL - side effect for printing a \code{data.frame} object
}
\description{
Show Specimen Method
}
