% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{show,specimens-method}
\alias{show,specimens-method}
\title{Show Specimens object as a dataframe}
\usage{
\S4method{show}{specimens}(object)
}
\arguments{
\item{object}{specimens object}
}
\value{
NULL - side effect for printing a \code{data.frame} object
}
\description{
Show Specimens object as a dataframe
}
