% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{show,speleothem-method}
\alias{show,speleothem-method}
\title{Show Speleothem Method}
\usage{
\S4method{show}{speleothem}(object)
}
\arguments{
\item{object}{speleothem object}
}
\value{
null - side effect, prints a \code{data.frame} with \code{speleo} metadata
}
\description{
Show Speleothem Method
}
