% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{show,speleothems-method}
\alias{show,speleothems-method}
\title{Show Datasets object as a dataframe}
\usage{
\S4method{show}{speleothems}(object)
}
\arguments{
\item{object}{datasets object}
}
\value{
null - side effect, prints a \code{data.frame} with \code{datasets} metadata
}
\description{
Show Datasets object as a dataframe
}
