% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{show,taxa-method}
\alias{show,taxa-method}
\title{Show Taxa Method}
\usage{
\S4method{show}{taxa}(object)
}
\arguments{
\item{object}{taxon object}
}
\value{
null - side effect, prints a \code{data.frame} with \code{taxon} metadata
}
\description{
Show Taxa Method
}
