% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{show,taxon-method}
\alias{show,taxon-method}
\title{Show Taxon Method}
\usage{
\S4method{show}{taxon}(object)
}
\arguments{
\item{object}{taxon object}
}
\value{
null - side effect, prints a \code{data.frame} with \code{taxon} metadata
}
\description{
Show Taxon Method
}
