% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{showDatasets,sites-method}
\alias{showDatasets,sites-method}
\title{Convert sites object to a \code{data.frame}}
\usage{
\S4method{showDatasets}{sites}(object)
}
\arguments{
\item{object}{A sites object}
}
\description{
Convert sites object to a \code{data.frame}
}
