% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{showDatasets}
\alias{showDatasets}
\title{Convert sites object to a \code{data.frame}}
\usage{
showDatasets(object)
}
\arguments{
\item{object}{A sites object}
}
\description{
Convert sites object to a \code{data.frame}
}
