% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{showMatch}
\alias{showMatch}
\title{Show matches for objects.}
\usage{
showMatch(x)
}
\arguments{
\item{x}{object to show matches for}
}
\value{
data.frame that marks if a \code{site} exists in another \code{sites} object
}
\description{
Show matches for objects.
}
