% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{sites-class}
\alias{sites-class}
\alias{sites}
\title{An S4 class for multi-site information from the Neotoma Paleoecology Database.}
\description{
An S4 class for multi-site information from the Neotoma Paleoecology Database.
}
