% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{sites_classes}
\alias{sites_classes}
\alias{site-class}
\alias{sites-class}
\title{An S4 class for \code{sites} information}
\value{
object of class \code{sites}
}
\description{
The S4 class for sites in the Neotoma Paleoecology Database.
}
