% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{specimen-class}
\alias{specimen-class}
\title{S4 class for specimen information}
\value{
object of class \code{specimen}
}
\description{
The standard object class for specimen
from the Neotoma Paleoecology Database.
}
