% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{specimens-class}
\alias{specimens-class}
\title{S4 class for specimens information}
\value{
object of class \code{specimens}
}
\description{
The grouped class for specimens from
the Neotoma Paleoecology Database.
}
