% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens.R
\name{specimens,collunit-method}
\alias{specimens,collunit-method}
\title{specimens}
\usage{
\S4method{specimens}{collunit}(x)
}
\arguments{
\item{x}{collunit object}
}
\description{
Obtain elements from collunit
}
