% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens.R
\name{specimens,collunits-method}
\alias{specimens,collunits-method}
\title{specimens}
\usage{
\S4method{specimens}{collunits}(x)
}
\arguments{
\item{x}{\code{collunits} object}
}
\value{
\code{data.frame} with specimens summary table
}
\description{
Obtain specimen elements from \code{collunits}
}
