% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens.R
\name{specimens,sites-method}
\alias{specimens,sites-method}
\title{specimens}
\usage{
\S4method{specimens}{sites}(x)
}
\arguments{
\item{x}{Use a sites element that has specimens added.}
}
\value{
The function returns a specimens summary table.
}
\description{
Information table for Specimens
}
\examples{
\dontrun{
# To return a specimens table do:
my_specimens <- get_specimens(19832)
my_tbl <- specimens(my_specimens)
}
}
\author{
Socorro Dominguez \email{s.dominguez@ht-data.com}
}
