% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{specimens}
\alias{specimens}
\alias{specimens-class}
\title{S4 class for \code{specimens} information}
\value{
object of class \code{specimens}
}
\description{
The standard object class for \code{specimens}
from the Neotoma Paleoecology Database.
}
