% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{speleothem-class}
\alias{speleothem-class}
\title{S4 class for speleothem information}
\value{
object of class \code{speleothem}
}
\description{
The standard object class for speleothem
from the Neotoma Paleoecology Database.
}
