% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothemdetails.R
\name{speleothemdetails,collunit-method}
\alias{speleothemdetails,collunit-method}
\title{speleothemdetails}
\usage{
\S4method{speleothemdetails}{collunit}(x)
}
\arguments{
\item{x}{collunit object}
}
\value{
\code{data.frame} with sample and speleothemdetails records
}
\description{
Obtain elements from collunit
}
