% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothemdetails.R
\name{speleothemdetails,collunits-method}
\alias{speleothemdetails,collunits-method}
\title{speleothemdetails}
\usage{
\S4method{speleothemdetails}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain elements from collunits
}
