% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{speleothemdetails}
\alias{speleothemdetails}
\title{Obtain speleothems and samples from a record or multiple records.}
\usage{
speleothemdetails(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at speleothem level
}
\description{
Obtain speleothems and samples from a record or multiple records.
}
