% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{speleothems-class}
\alias{speleothems-class}
\title{Speleothem}
\value{
object of class \code{speleothem}
}
\description{
Speleothem class for SISAL data
from the Neotoma Paleoecology Database.
}
