% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothems.R
\name{speleothems,collunit-method}
\alias{speleothems,collunit-method}
\title{speleothems}
\usage{
\S4method{speleothems}{collunit}(x)
}
\arguments{
\item{x}{collunit object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain elements from collunit
}
