% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothems.R
\name{speleothems,collunits-method}
\alias{speleothems,collunits-method}
\title{Get speleothems from a collectionunit or set of collection units:}
\usage{
\S4method{speleothems}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain elements from collunits
}
