% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothems.R
\name{speleothems,site-method}
\alias{speleothems,site-method}
\title{speleothems}
\usage{
\S4method{speleothems}{site}(x)
}
\arguments{
\item{x}{site object}
}
\value{
\code{data.frame} with speleothem records
}
\description{
Obtain elements on the speleothems level
}
\examples{
\donttest{
kesang <- get_sites(sitename = "Kesang cave") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_speleothems()
sp <- speleothems(kesang)
}
}
