% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R,
%   R/02_genericDefinitions.R, R/speleothems.R
\docType{class}
\name{speleothems}
\alias{speleothems}
\alias{speleothems-class}
\alias{speleothems,site-method}
\alias{speleothems,collunits-method}
\alias{speleothems,collunit-method}
\title{S4 class for \code{speleothem} information}
\usage{
speleothems(x)

\S4method{speleothems}{site}(x)

\S4method{speleothems}{collunits}(x)

\S4method{speleothems}{collunit}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
object of class \code{speleothems}

data.frame with record information at speleothem level
}
\description{
The S4 class for \code{speleothem} data.
}
