% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{speleothems_classes}
\alias{speleothems_classes}
\alias{speleothem-class}
\alias{speleothems-class}
\title{S4 class for \code{speleothem} information}
\value{
object of class \code{speleothems}
}
\description{
The S4 class for \code{speleothem} data.
}
