% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{[,collunits,numeric,ANY,ANY-method}
\alias{[,collunits,numeric,ANY,ANY-method}
\title{Get or remove sites by numeric index}
\usage{
\S4method{[}{collunits,numeric,ANY,ANY}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{The collunits object}

\item{i}{The numeric index}
}
\description{
Get or remove sites by numeric index
}
