% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{[,collunits,numeric-method}
\alias{[,collunits,numeric-method}
\title{Get or remove sites by numeric index}
\usage{
\S4method{[}{collunits,numeric}(x, i)
}
\arguments{
\item{x}{The collunits object}

\item{i}{The numeric index}
}
\value{
null used for side effects. Printing a data.frame
}
\description{
Retrieve sites by numeric index
}
