% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{[,datasets,numeric,ANY,ANY-method}
\alias{[,datasets,numeric,ANY,ANY-method}
\title{Get or remove datasets by numeric index}
\usage{
\S4method{[}{datasets,numeric,ANY,ANY}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{The datasets object}

\item{i}{The numeric index}
}
\description{
Get or remove datasets by numeric index
}
