% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{[,datasets,numeric,ANY-method}
\alias{[,datasets,numeric,ANY-method}
\title{Get or remove datasets by numeric index}
\usage{
\S4method{[}{datasets,numeric,ANY}(x, i)
}
\arguments{
\item{x}{The datasets object}

\item{i}{The numeric index}
}
\value{
Get or remove \code{datasets} by numeric index
}
\description{
Get or remove datasets by numeric index
}
