% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[,site,character,ANY-method}
\alias{[,site,character,ANY-method}
\title{Get site field by character index}
\usage{
\S4method{[}{site,character,ANY}(x, i)
}
\arguments{
\item{x}{The site object}

\item{i}{The column indicator}
}
\value{
sliced \code{site} object
}
\description{
Get site field by character index
}
