% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[,sites,numeric,ANY-method}
\alias{[,sites,numeric,ANY-method}
\title{Get or remove sites by numeric index}
\usage{
\S4method{[}{sites,numeric,ANY}(x, i)
}
\arguments{
\item{x}{The sites object}

\item{i}{The numeric index}
}
\description{
Get or remove sites by numeric index
}
\examples{
some_site <- get_sites(sitename = "Site\%")
some_site[2]
}
