% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[,sites,numeric-method}
\alias{[,sites,numeric-method}
\title{Get or remove sites by numeric index}
\usage{
\S4method{[}{sites,numeric}(x, i)
}
\arguments{
\item{x}{The sites object}

\item{i}{The numeric index}
}
\value{
sliced \code{site} object
}
\description{
Get or remove sites by numeric index
}
