% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{[,specimens,numeric,ANY-method}
\alias{[,specimens,numeric,ANY-method}
\title{Get or remove specimens by numeric index}
\usage{
\S4method{[}{specimens,numeric,ANY}(x, i)
}
\arguments{
\item{x}{The specimens object}

\item{i}{The numeric index}
}
\value{
sliced \code{specimens} object
}
\description{
Get or remove specimens by numeric index
}
