% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{[,speleothems,numeric-method}
\alias{[,speleothems,numeric-method}
\title{Get or remove speleothems by numeric index}
\usage{
\S4method{[}{speleothems,numeric}(x, i)
}
\arguments{
\item{x}{The speleothems object}

\item{i}{The numeric index}
}
\value{
Get or remove \code{speleothems} by numeric index
}
\description{
Get or remove speleothems by numeric index
}
