% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{[[,chronologies,numeric,ANY-method}
\alias{[[,chronologies,numeric,ANY-method}
\title{Slicer}
\usage{
\S4method{[[}{chronologies,numeric,ANY}(x, i)
}
\arguments{
\item{x}{chronologies object}

\item{i}{iteration in chronologies list}
}
\description{
Obtain one of the elements within a chronologies list
}
