% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{[[,chronologies,numeric-method}
\alias{[[,chronologies,numeric-method}
\title{Extract}
\usage{
\S4method{[[}{chronologies,numeric}(x, i)
}
\arguments{
\item{x}{chronologies object}

\item{i}{iteration in chronologies list}
}
\value{
selected \code{chronology} object
}
\description{
Obtain one of the elements within a \code{chronologies} list
either by element order or by element name.
}
