% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{[[,collunits,numeric-method}
\alias{[[,collunits,numeric-method}
\title{Slicer}
\usage{
\S4method{[[}{collunits,numeric}(x, i)
}
\arguments{
\item{x}{collunits object}

\item{i}{iteration in collunits list}
}
\value{
sliced \code{collunits} object
}
\description{
Obtain one of the elements within a collunits list
}
