% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{[[,contacts,numeric,ANY-method}
\alias{[[,contacts,numeric,ANY-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[[}{contacts,numeric,ANY}(x, i)
}
\arguments{
\item{x}{A contact object.}

\item{i}{The numeric index of a contact slot.}
}
\value{
sliced contacts
}
\description{
Extract or Replace Parts of an Object
}
