% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{[[,datasets,numeric,ANY-method}
\alias{[[,datasets,numeric,ANY-method}
\title{Obtain one of the elements within a datasets list}
\usage{
\S4method{[[}{datasets,numeric,ANY}(x, i)
}
\arguments{
\item{x}{datasets object}

\item{i}{iteration in datasets list}
}
\description{
in progress
}
