% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{[[,publications,numeric,ANY-method}
\alias{[[,publications,numeric,ANY-method}
\title{Obtain one of the elements within a publication list.}
\usage{
\S4method{[[}{publications,numeric,ANY}(x, i)
}
\arguments{
\item{x}{A publications object.}

\item{i}{A numeric index for the requested publication}
}
\value{
selected \code{publications} object from index
}
\description{
Obtain one of the elements within a publication list.
}
