% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{[[,sites,numeric,ANY-method}
\alias{[[,sites,numeric,ANY-method}
\title{Obtain one of the elements within a sites list}
\usage{
\S4method{[[}{sites,numeric,ANY}(x, i)
}
\arguments{
\item{x}{sites object}

\item{i}{iteration in sites list}
}
\description{
in progress
}
