% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{[[,specimens,numeric-method}
\alias{[[,specimens,numeric-method}
\title{Slicer}
\usage{
\S4method{[[}{specimens,numeric}(x, i)
}
\arguments{
\item{x}{specimens object}

\item{i}{iteration in specimens list}
}
\value{
sliced \code{specimens} object
}
\description{
Obtain one of the elements within a specimens list
}
