% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{[[,speleothems,numeric-method}
\alias{[[,speleothems,numeric-method}
\title{Slicer}
\usage{
\S4method{[[}{speleothems,numeric}(x, i)
}
\arguments{
\item{x}{speleothems object}

\item{i}{iteration in speleothems list}
}
\value{
sliced \code{dataset} object
}
\description{
Obtain one of the elements within a datasets list
}
