% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{[[,taxa,numeric-method}
\alias{[[,taxa,numeric-method}
\title{Slicer}
\usage{
\S4method{[[}{taxa,numeric}(x, i)
}
\arguments{
\item{x}{taxa object}

\item{i}{iteration in taxa list}
}
\value{
sliced \code{taxa} object
}
\description{
Obtain one of the elements within a taxa list
}
