% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdataset-methods.R
\name{[[<-,datasets,ANY,ANY,ANY-method}
\alias{[[<-,datasets,ANY,ANY,ANY-method}
\title{Insert dataset}
\usage{
\S4method{[[}{datasets,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{datasets object}

\item{i}{iteration in datasets list}

\item{j}{}
}
\description{
Obtain one of the elements within a datasets list
}
