% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{[[<-,publications,ANY,ANY,ANY-method}
\alias{[[<-,publications,ANY,ANY,ANY-method}
\title{Assign value to an element in a publication list.}
\usage{
\S4method{[[}{publications,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{A publications object.}

\item{i}{A numeric index for the requested publication}

\item{value}{The value to be used}
}
\value{
\code{publications} with new assigned value.
}
\description{
Assign value to an element in a publication list.
}
