% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{[[<-,samples,ANY,ANY,ANY-method}
\alias{[[<-,samples,ANY,ANY,ANY-method}
\title{Insert sample}
\usage{
\S4method{[[}{samples,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{samples object}

\item{i}{iteration in samples list}

\item{value}{The value to be used}
}
\value{
\code{samples} object with reassigned values
}
\description{
Obtain one of the elements within a samples list
}
