% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[[<-,sites,ANY,ANY,ANY-method}
\alias{[[<-,sites,ANY,ANY,ANY-method}
\title{Insert site}
\usage{
\S4method{[[}{sites,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{sites object}

\item{i}{iteration in sites list}

\item{j}{}
}
\description{
Obtain one of the elements within a sites list
}
