% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{[[<-,specimens,ANY,ANY,ANY-method}
\alias{[[<-,specimens,ANY,ANY,ANY-method}
\title{Insert specimen}
\usage{
\S4method{[[}{specimens,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{specimens object}

\item{i}{iteration in specimens list}

\item{value}{The value to be used}
}
\value{
\code{specimens} object with reassigned values
}
\description{
Obtain one of the elements within a specimens list
}
