% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{[[<-,speleothems,ANY,ANY,ANY-method}
\alias{[[<-,speleothems,ANY,ANY,ANY-method}
\title{Insert dataset}
\usage{
\S4method{[[}{speleothems,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{speleothems object}

\item{i}{iteration in speleothems list}

\item{value}{The value to be used}
}
\value{
One \code{speleothem} slot's value
}
\description{
Obtain one of the elements within a speleothems list
}
