% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{[[<-,taxa,ANY,ANY,ANY-method}
\alias{[[<-,taxa,ANY,ANY,ANY-method}
\title{Insert taxon}
\usage{
\S4method{[[}{taxa,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{taxa object}

\item{i}{iteration in taxa list}

\item{value}{The value to be used}
}
\value{
One \code{taxon} slot's value
}
\description{
Obtain one of the elements within a taxa list
}
