% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{[,taxa,numeric-method}
\alias{[,taxa,numeric-method}
\title{Get or remove taxa by numeric index}
\usage{
\S4method{[}{taxa,numeric}(x, i)
}
\arguments{
\item{x}{The taxa object}

\item{i}{The numeric index}
}
\value{
Get or remove \code{taxa} by numeric index
}
\description{
Get or remove taxa by numeric index
}
