% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{[<-,dataset,character,ANY,ANY-method}
\alias{[<-,dataset,character,ANY,ANY-method}
\title{Assign dataset field by numeric index}
\usage{
\S4method{[}{dataset,character,ANY,ANY}(x, i, j, drop = FALSE) <- value
}
\arguments{
\item{x}{The dataset object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\description{
Assign dataset field by numeric index
}
