% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{[<-,sample,character,ANY,ANY-method}
\alias{[<-,sample,character,ANY,ANY-method}
\title{Assign sample field by numeric index}
\usage{
\S4method{[}{sample,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The sample object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{sample} object with reassigned character values
}
\description{
Assign sample field by numeric index
}
