% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[<-,site,character,ANY,ANY-method}
\alias{[<-,site,character,ANY,ANY-method}
\title{Assign site field by numeric index}
\usage{
\S4method{[}{site,character,ANY,ANY}(x, i, j, drop = FALSE) <- value
}
\arguments{
\item{x}{The site object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\description{
Assign site field by numeric index
}
