% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[<-,site,numeric,ANY,ANY-method}
\alias{[<-,site,numeric,ANY,ANY-method}
\title{Assign site field by numeric index}
\usage{
\S4method{[}{site,numeric,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The site object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{sites} object with reassigned numeric values
}
\description{
Assign site field by numeric index
}
