% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{[<-,specimen,character,ANY,ANY-method}
\alias{[<-,specimen,character,ANY,ANY-method}
\title{Assign specimen field by numeric index}
\usage{
\S4method{[}{specimen,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The specimen object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{specimen} object with reassigned character values
}
\description{
Assign specimen field by numeric index
}
