% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothem-methods.R
\name{[<-,speleothem,character,ANY,ANY-method}
\alias{[<-,speleothem,character,ANY,ANY-method}
\title{Assign speleothem field by numeric index}
\usage{
\S4method{[}{speleothem,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The speleothem object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{speleothem} slot with new assigned character value
}
\description{
Assign speleothem field by numeric index
}
