% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-methods.R
\name{[<-,taxon,numeric,ANY,ANY-method}
\alias{[<-,taxon,numeric,ANY,ANY-method}
\title{Assign taxon field by numeric index}
\usage{
\S4method{[}{taxon,numeric,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The taxon object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{taxon} slot with new assigned numeric value
}
\description{
Assign taxon field by numeric index
}
