% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{taxa-class}
\alias{taxa-class}
\title{S4 class for taxa information}
\value{
object of class \code{taxon}
}
\description{
Taxa class for taxa information
from the Neotoma Paleoecology Database.
}
