% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa,collunits-method}
\alias{taxa,collunits-method}
\title{Extract taxonomic data from a set of sites.}
\usage{
\S4method{taxa}{collunits}(object)
}
\arguments{
\item{object}{A \code{collunits} object.}
}
\value{
A \code{data.frame} reporting the taxa/data objects, units, &cetera within the set of records.
}
\description{
Extract taxonomic data from a set of sites.
}
