% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R,
%   R/02_genericDefinitions.R, R/taxa.R
\docType{class}
\name{taxa}
\alias{taxa}
\alias{taxa-class}
\alias{taxa,site-method}
\alias{taxa,collunits-method}
\alias{taxa,collunit-method}
\title{S4 class for taxa information}
\usage{
taxa(object)

\S4method{taxa}{site}(object)

\S4method{taxa}{collunits}(object)

\S4method{taxa}{collunit}(object)
}
\arguments{
\item{object}{Sites object to extract taxa table from}
}
\value{
object of class \code{taxon}

data.frame with taxa records
}
\description{
Taxa details from the Neotoma Paleoecology Database.

Show the samples table
}
