% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{taxa_classes}
\alias{taxa_classes}
\alias{taxon-class}
\alias{taxa-class}
\title{S4 class for taxa information}
\value{
object of class \code{taxon}
}
\description{
Taxa details from the Neotoma Paleoecology Database.
}
