% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{taxon-class}
\alias{taxon-class}
\title{S4 class for specimen information}
\value{
object of class \code{taxon}
}
\description{
Taxon class for single taxon information
from the Neotoma Paleoecology Database.
}
