% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNull.R
\name{testNull}
\alias{testNull}
\title{Clear out NULL values and replace with value}
\usage{
testNull(val, out = NA)
}
\arguments{
\item{val}{A value to be passed in}

\item{out}{A default value to be returned if \code{val} is NULL.}
}
\value{
The values passed in \code{val} or \code{out} if \code{val} is \code{NULL}.
}
\description{
Similar to concatenate, checks to see if a value is NULL
and replace it with a known value.
}
