% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJSON.R
\name{toJSON,sites-method}
\alias{toJSON,sites-method}
\title{toJSON}
\usage{
\S4method{toJSON}{sites}(x = NA)
}
\arguments{
\item{x}{sites R object to be converted}
}
\value{
The function returns a character string in JSON format
}
\description{
Convert a Neotoma2 \code{sites} object into a standardized JSON file for API management.
}
\examples{
\dontrun{
# To find all sites that contain the string "Alexander\%"
alex.sites <- get_sites(sitename="Alexander\%")
# Convert the object to json
toJSON(alex.sites)
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
