% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_na.R
\name{use_na}
\alias{use_na}
\title{Change NA values from logic to a prescribed type.}
\usage{
use_na(x, type)
}
\arguments{
\item{x}{An element that may or may not have NA values.}

\item{type}{A character string with values either \code{char} or \code{int}.}
}
\value{
object converted to \code{NA_character} or \code{NA_integer}
}
\description{
Pass an object and convert all
\code{NA} elements to particular \code{NA} types.
}
