% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{write.csv,datasets-method}
\alias{write.csv,datasets-method}
\title{write CSV}
\usage{
\S4method{write.csv}{datasets}(x, ...)
}
\arguments{
\item{x}{datasets object}

\item{...}{Additional parameters associated with the call.}
}
\value{
null -side effect for printing a CSV file
}
\description{
write CSV
}
