% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{write.csv,sites-method}
\alias{write.csv,sites-method}
\title{write CSV}
\usage{
\S4method{write.csv}{sites}(x, ...)
}
\arguments{
\item{x}{A sites object}

\item{...}{Other options to pass to \code{write.csv()}.}
}
\value{
NULL side effect from saving a csv file
}
\description{
write CSV
}
