% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{write.csv,specimens-method}
\alias{write.csv,specimens-method}
\title{write CSV}
\usage{
\S4method{write.csv}{specimens}(x, ...)
}
\arguments{
\item{x}{specimens object}

\item{...}{Additional parameters associated with the call.}
}
\value{
NULL - side effect of saving CSV file
}
\description{
write CSV
}
