% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{add_chronology,collunit,chronology,data.frame-method}
\alias{add_chronology,collunit,chronology,data.frame-method}
\title{Add a new chronology into an existing collectionunit.}
\usage{
\S4method{add_chronology}{collunit,chronology,data.frame}(object, x, y)
}
\arguments{
\item{object}{A collection unit object}

\item{x}{A chronology object generated using \code{set_chronology()}}

\item{y}{A data.frame of sample ages, with required columns:
"analysisunitid", "age", "agetype", "ageolder", and "ageyounger".}
}
\description{
Add a new chronology into an existing collectionunit.
}
