% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{as.data.frame,chronologies-method}
\alias{as.data.frame,chronologies-method}
\title{as.data.frame chronologies}
\usage{
\S4method{as.data.frame}{chronologies}(x)
}
\arguments{
\item{x}{chronologies object}
}
\value{
data.frame with chronologies metadata
}
\description{
Convert all slots within each chronology within a
chronologies object to a data.frame.
}
