% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{as.data.frame,chronology-method}
\alias{as.data.frame,chronology-method}
\title{Create a data.frame from a chronology object.}
\usage{
\S4method{as.data.frame}{chronology}(x)
}
\arguments{
\item{x}{chronology object}
}
\value{
data.frame
}
\description{
Convert all slots within a chronology to a data.frame.
}
