% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{as.data.frame,specimens-method}
\alias{as.data.frame,specimens-method}
\title{as.data.frame specimens}
\usage{
\S4method{as.data.frame}{specimens}(x)
}
\arguments{
\item{x}{specimens object}
}
\value{
\code{data.frame} with \verb{specimens metadata}
}
\description{
show as data.frame
}
