% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{authors-class}
\alias{authors-class}
\title{An S4 class for a set of Neotoma author objects.}
\value{
object of class \code{authors}
}
\description{
The S4 \code{authors} are a set of individual \code{author} objects that
are then associated with a single S4 \code{publication} class.
}
\examples{
{
simon <- new("contact", familyname = "Goring", givennames = "Simon J.")
socorro <- new("contact", familyname = "Dominguez", givennames = "Socorro")
first_author <- new("author", author = simon, order = 1)
second_author <- new("author", author = socorro, order = 2)
author_list <- new("authors", authors = list(first_author, second_author))
}
}
