% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_chron.R
\name{build_chron}
\alias{build_chron}
\title{build_chron}
\usage{
build_chron(x)
}
\arguments{
\item{x}{A chronology element from the API JSON output.}
}
\value{
A single \code{chronology} object.
}
\description{
A helper function to build a new chronology object from the
Neotoma API response.
}
\details{
This function is an internal function called from
\code{build_collunit()} to help support the translation between the JSON
representation of data in the API and the R implementation.
}
\author{
Socorro Dominguez
}
