% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataset.R
\name{build_dataset}
\alias{build_dataset}
\title{get_downloads}
\usage{
build_dataset(x)
}
\arguments{
\item{x}{dataset list}
}
\value{
list parsed into datasets
\code{"try-error"} describing the reason for failure
(either mis-defined parameters or an error from the Neotoma API),
or a table of sites, with rows corresponding to the number of
individual sites and datasets returned by the Neotoma API.
Each "dataset" contains the
following detail of information:
\item{ \code{datasetid} }{dataset ID number}
\item{ \code{datasetname} }{site's name}
\item{ \code{datasettype} }{type of data found}
\item{ \code{location} }{sf object that describes site's location}
\item{ \code{notes} }{notes on the dataset}
\item{ \code{taxa table} }{taxa table}
\item{ \code{pi list} }{P.I. info}
\item{ \code{analyst} }{analyst info}
\item{ \code{metadata} }{dataset metadata}
}
\description{
Helper function to build a dataset
}
\examples{
\dontrun{
# To build dataset from API call:
build_dataset(x)
}

}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
