% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataset.R
\name{build_dataset}
\alias{build_dataset}
\title{Build a dataset object from a JSON list representation.}
\usage{
build_dataset(x)
}
\arguments{
\item{x}{a JSON dataset object passed from the Neotoma API.}
}
\value{
A simple \code{dataset} object.
}
\description{
Helper function to build a dataset from the API JSON response.
}
\author{
Socorro Dominguez
}
