% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_sites}
\alias{build_sites}
\title{Build a \code{site} from the Neotoma API response.}
\usage{
build_sites(x)
}
\arguments{
\item{x}{A list returned from the Neotoma API \code{data} section.}
}
\value{
A simple \code{site} object
}
\description{
Build a \code{site} from the Neotoma API response.
}
