% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_specimen.R
\name{build_specimen}
\alias{build_specimen}
\title{get_downloads}
\usage{
build_specimen(x)
}
\arguments{
\item{x}{specimen list}
}
\value{
list parsed into specimen
}
\description{
Helper function to build a specimen
}
\examples{
\dontrun{
# To build dataset from API call:
build_specimen(x)
}

}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
