% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_specimen.R
\name{build_specimen}
\alias{build_specimen}
\title{Build a specimen objects.}
\usage{
build_specimen(x)
}
\arguments{
\item{x}{specimen list}
}
\value{
A simple \code{specimen} object
}
\description{
A helper function to build a specimen object from a list returned by
the Neotoma API call. The function is not exported, but called from
the get_specimens() call.
}
