% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/neotoma-methods.R
\name{c,missingOrNULL-method}
\alias{c,missingOrNULL-method}
\title{c Method for NULL values}
\usage{
\S4method{c}{missingOrNULL}(x = "missingORNULL", y)

\S4method{c}{missingOrNULL}(x = "missingORNULL", y)
}
\arguments{
\item{x}{NULL object}

\item{y}{sites/datasets object}
}
\value{
concatenated \code{collunits} object

\code{list} of concatenated items when the first object is NULL
}
\description{
c Method for NULL values

c Method for NULL values
}
