% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/contact-methods.R, R/dataset-methods.R, R/publication-methods.R,
%   R/samples-methods.R, R/site-methods.R, R/speleothem-methods.R,
%   R/taxon-methods.R
\docType{class}
\name{c,chronologies-method}
\alias{c,chronologies-method}
\alias{c,collunits-method}
\alias{c,contacts-method}
\alias{c,contact-method}
\alias{c,datasets-method}
\alias{c,publications-method}
\alias{c,samples-method}
\alias{c}
\alias{c,missingOrNULL-method}
\alias{c,sites-method}
\alias{c,speleothems-method}
\alias{c,taxa-method}
\title{c - Combine \code{neotoma2} objects}
\usage{
\S4method{c}{chronologies}(x, y)

\S4method{c}{collunits}(x, y)

\S4method{c}{contacts}(x, y)

\S4method{c}{contact}(x, y)

\S4method{c}{datasets}(x, y)

\S4method{c}{publications}(x, y)

\S4method{c}{samples}(x, y)

\S4method{c}{missingOrNULL}(x = "missingORNULL", y)

\S4method{c}{missingOrNULL}(x = "missingORNULL", y)

\S4method{c}{sites}(x, y)

\S4method{c}{speleothems}(x, y)

\S4method{c}{taxa}(x, y)
}
\arguments{
\item{x}{\code{neotoma} object 1 or NULL}

\item{y}{\code{neotoma} object 2 or NULL}
}
\value{
concatenated and cleaned \code{sites} object
}
\description{
c - Combine \code{neotoma2} objects
}
