% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$,chronologies-method}
\alias{$,chronologies-method}
\title{$ for chronologies}
\usage{
\S4method{$}{chronologies}(x, name)
}
\arguments{
\item{x}{chronologies object}

\item{name}{name of the slot}
}
\description{
Obtain slots of a chronologies object without using at-mark
}
