% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$,chronologies-method}
\alias{$,chronologies-method}
\title{Extract}
\usage{
\S4method{$}{chronologies}(x, name)
}
\arguments{
\item{x}{chronologies object}

\item{name}{name of the slot}
}
\value{
A multiple \code{chronologies}
}
\description{
Obtain chronology slots across all chronology elements within
a chronologies object.
}
