% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$,chronology-method}
\alias{$,chronology-method}
\title{$}
\usage{
\S4method{$}{chronology}(x, name)
}
\arguments{
\item{x}{chronology object}
}
\description{
Obtain slots of a chronology without using at-mark
}
