% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$,chronology-method}
\alias{$,chronology-method}
\title{Extract}
\usage{
\S4method{$}{chronology}(x, name)
}
\arguments{
\item{x}{chronology object}

\item{name}{name of the slot}
}
\value{
value of the \code{slot} name
}
\description{
Extract chronology metadata by slot name.
}
