% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{$,collunits-method}
\alias{$,collunits-method}
\title{$ for collunits}
\usage{
\S4method{$}{collunits}(x, name)
}
\arguments{
\item{x}{collunits object}

\item{name}{name of the slot}
}
\value{
null prints element of a slot
}
\description{
Obtain slots of a collunit without using at-mark
}
