% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{$,datasets-method}
\alias{$,datasets-method}
\title{$ for datasets}
\usage{
\S4method{$}{datasets}(x, name)
}
\arguments{
\item{x}{datasets object}

\item{name}{name of the slot.}
}
\value{
Obtain a \code{datasets}' \code{slot} value using $
}
\description{
Obtain slots of a dataset without using at-mark
}
