% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{$,sample-method}
\alias{$,sample-method}
\title{$}
\usage{
\S4method{$}{sample}(x, name)
}
\arguments{
\item{x}{sample object}

\item{name}{name of the slot}
}
\value{
\code{value} at selected slot
}
\description{
Obtain slots of a sample without using at-mark
}
