% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{$,samples-method}
\alias{$,samples-method}
\title{$ for samples}
\usage{
\S4method{$}{samples}(x, name)
}
\arguments{
\item{x}{samples object}

\item{name}{name of the slot}
}
\value{
\code{value} at selected slot
}
\description{
Obtain slots of a site without using at-mark
}
