% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{$<-,chronology-method}
\alias{$<-,chronology-method}
\title{Replace part of an object}
\usage{
\S4method{$}{chronology}(x, name) <- value
}
\arguments{
\item{x}{A chronology object}

\item{name}{The name of the chronology slot.}

\item{value}{A value to be assigned to the chronology slot.}
}
\value{
reassigned \code{chronology} object
}
\description{
Assign values to slots within a chronology object.
}
