% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{$<-,collunit-method}
\alias{$<-,collunit-method}
\title{Assign collunit field by numeric index}
\usage{
\S4method{$}{collunit}(x, name) <- value
}
\arguments{
\item{x}{The collunit object.}

\item{name}{name of the slot.}

\item{value}{The value to be used.}
}
\value{
assign a new value to a slot
}
\description{
Assign collunit field by numeric index
}
