% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{$<-,dataset-method}
\alias{$<-,dataset-method}
\title{Assign dataset field by numeric index}
\usage{
\S4method{$}{dataset}(x, name) <- value
}
\arguments{
\item{x}{The dataset object.}

\item{name}{name of the slot.}

\item{value}{The value to be used.}
}
\value{
Assign new \code{dataset} by numeric index
}
\description{
Assign dataset field by numeric index
}
