% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/dataset-methods.R, R/site-methods.R, R/speleothem-methods.R,
%   R/taxon-methods.R
\name{$<-,chronology-method}
\alias{$<-,chronology-method}
\alias{$<-,collunit-method}
\alias{cash-set,collunit-method}
\alias{$<-,dataset-method}
\alias{cash-set,dataset-method}
\alias{cash-set}
\alias{$<-,site-method}
\alias{$<-,speleothem-method}
\alias{cash-set,speleothem-method}
\alias{$<-,taxon-method}
\alias{cash-set,taxon-method}
\title{cash-set}
\usage{
\S4method{$}{chronology}(x, name) <- value

\S4method{$}{collunit}(x, name) <- value

\S4method{$}{dataset}(x, name) <- value

\S4method{$}{site}(x, name) <- value

\S4method{$}{speleothem}(x, name) <- value

\S4method{$}{taxon}(x, name) <- value
}
\arguments{
\item{x}{A \code{neotoma2} object.}

\item{name}{name of the slot}

\item{value}{The value to be used.}
}
\value{
\code{neotoma2} object with reassigned values
}
\description{
cash-set
}
