% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen-methods.R
\name{$,specimen-method}
\alias{$,specimen-method}
\title{$}
\usage{
\S4method{$}{specimen}(x, name)
}
\arguments{
\item{x}{specimen object}

\item{name}{name of the slot}
}
\value{
value at chosen slot in the \code{specimen} object
}
\description{
Obtain slots of a specimen without using at-mark
}
