% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args}
\alias{check_args}
\title{check_args}
\usage{
check_args(cl)
}
\arguments{
\item{cl}{called arguments.
Arguments are going to be called by match_call inside:
\code{\link{get_sites}}
\code{\link{get_datasets}}
\code{\link{get_downloads}}}
}
\value{
A list with two components:
\item{flag}{Returns a 0 if everything's fine, a 1 if there's a problem.}
\item{message}{A list of error messages.}
}
\description{
Internal function to check passed arguments.
}
\references{
Neotoma Project Website: https://www.neotomadb.org/
}
\author{
Socorro Dominguez
}
