% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contacts.R
\name{check_contacts.contacts}
\alias{check_contacts.contacts}
\title{Get contact information for Neotoma contributors}
\usage{
\method{check_contacts}{contacts}(x, similarity = 0.5, ...)
}
\arguments{
\item{x}{contacts A \code{contacts} object associated with a set of names.}

\item{similarity}{The similarity score between matched records (from 0 - 1).}

\item{...}{Additional parameters associated with the call.}
}
\value{
contacts object
}
\description{
Get contact information for Neotoma contributors
}
