% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contacts.R
\name{check_contacts}
\alias{check_contacts}
\title{Check contact information for a record against Neotoma contributors}
\usage{
check_contacts(x)
}
\arguments{
\item{x}{contacts A \code{contacts} object associated with a set of names.}

\item{contactname}{A full or partial name for an individual contributor to the database.}

\item{familyname}{The full or partial last name for an individual contributor to the database.}

\item{status}{The current status of the contributor (\code{active} or \code{retired})}
}
\description{
Uses the Neotoma API to search and access information about individuals who have contributed to the data in the Neotoma Paleoecology Database
}
