% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,site-method}
\alias{chroncontrols,site-method}
\title{Recover information about the chron controls for a collectionunit.}
\usage{
\S4method{chroncontrols}{site}(x)
}
\arguments{
\item{x}{site object}
}
\value{
data.frame with chronological controls
}
\description{
For a site that includes collection units with chronologies
return the chronological controls that are used in building the chronology.
}
