% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroncontrols.R
\name{chroncontrols,sites-method}
\alias{chroncontrols,sites-method}
\title{Recover information about the chron controls for a collectionunit.}
\usage{
\S4method{chroncontrols}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with chronological controls
}
\description{
For all sites that includes collection units with chronologies
return the chronological controls that are used in building the chronology.
}
