% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{chroncontrols}
\alias{chroncontrols}
\title{chroncontrols}
\usage{
chroncontrols(x)
}
\arguments{
\item{x}{Sites object to extract chroncontrols table from}
}
\value{
data.frame with chroncontrols information
}
\description{
Show the samples table
}
