% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{chronologies-class}
\alias{chronologies-class}
\title{S4 class for chronologies information}
\value{
object of class \code{chronologies}
}
\description{
The grouped class for chronologies
from the Neotoma Paleoecology Database.
}
