% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{chronologies,collunit-method}
\alias{chronologies,collunit-method}
\title{Extract \code{chronologies} from a \code{collunit} object.}
\usage{
\S4method{chronologies}{collunit}(x)
}
\arguments{
\item{x}{A \code{collunit} object}
}
\value{
\code{chronologies} from a \code{collunit} object
}
\description{
Extract \code{chronologies} from a \code{collunit} object.
}
