% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/subset-methods.R
\name{chronologies}
\alias{chronologies}
\alias{chronologies,collunits-method}
\alias{chronologies,site-method}
\alias{chronologies,sites-method}
\title{Obtain the chronology from a record or multiple records.}
\usage{
chronologies(x)

\S4method{chronologies}{collunits}(x)

\S4method{chronologies}{site}(x)

\S4method{chronologies}{sites}(x)
}
\arguments{
\item{x}{sites object that contains chronologies}
}
\value{
chronologies object with all chronologies used.
}
\description{
Obtain the chronology from a record or multiple records.
}
