% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{chronologies}
\alias{chronologies}
\title{Obtain the chronology from a record or multiple records.}
\usage{
chronologies(x)
}
\arguments{
\item{x}{sites object that contains chronologies}
}
\value{
chronologies object with all chronologies used.
}
\description{
Obtain the chronology from a record or multiple records.
}
