% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{chronology-class}
\alias{chronology-class}
\title{S4 class for chronologies information}
\value{
object of class \code{chronology}
}
\description{
The class for chronologies from the
Neotoma Paleoecology Database. A single collection unit may
have one or more chronology. These individual chronology
classes are then grouped into an S4 \code{chronologies} class.
}
