% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{cite_data,sites-method}
\alias{cite_data,sites-method}
\title{Obtain data citations from multiple records.}
\usage{
\S4method{cite_data}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
\code{data.frame} object with citation information.
}
\description{
Given complete dataset objects in Neotoma (must have used
\code{get_datasets()} or \code{get_downloads()}), return a formatted
citation for the record, including the dataset DOI.
}
\examples{
{
ds <- get_datasets(1)
cite_data(ds)
}
}
