% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{cite_data}
\alias{cite_data}
\title{Obtain the DOI for publications or datasets.}
\usage{
cite_data(x)
}
\arguments{
\item{x}{Object with DOIs associated to it.}
}
\description{
Obtain the DOI for publications or datasets.
}
