% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(x = NA, verbose = TRUE, ...)
}
\arguments{
\item{x}{sites, datasets, collunits that may have duplicates.}

\item{verbose}{parameter to prints out progress bar}

\item{...}{Additional parameters associated with the call.}
}
\description{
Function that removes duplicate objects such as sites, datasets,
or collection units.
}
\author{
Simon Goring \email{goring@wisc.edu}
}
