% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/neotoma-methods.R
\docType{class}
\name{collunit-class}
\alias{collunit-class}
\alias{collunit}
\title{S4 class for collunits information}
\description{
S4 class for collunits information

S4 class for collunits information
}
