% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{collunit-class}
\alias{collunit-class}
\title{S4 class for collection units information.}
\value{
object of class \code{collunit}
}
\description{
A collection unit represents a collection event from within
a site. For example, a lake sediment core, or a single dig site within an
archaeological site.
}
