% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{collunits-class}
\alias{collunits-class}
\title{An S4 class for Neotoma Collection Units}
\description{
Holds Collection unit information
from the Neotoma Paleoecology Database.
@returns object of class \code{collunits}
}
