% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{collunits,sites-method}
\alias{collunits,sites-method}
\title{Extract \code{collunits} from a \code{sites} object.}
\usage{
\S4method{collunits}{sites}(object)
}
\arguments{
\item{object}{A \code{sites} object}
}
\value{
\code{collunits} from a \code{sites} object
}
\description{
Extract \code{collunits} from a \code{sites} object.
}
