% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{collunits}
\alias{collunits}
\title{Extract collection units from a sites object}
\usage{
collunits(object)
}
\arguments{
\item{object}{A sites object}
}
\value{
collunits detail from a sites object
}
\description{
Extract collection units from a sites object
}
