% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{contact-class}
\alias{contact-class}
\title{An S4 class for Neotoma contacts}
\value{
object of class \code{contact}
}
\description{
The object that contains the contact information for an
individual, along with associated metadata.
}
\examples{
new("contact", familyname = "Goring", givennames = "Simon J.")
}
