% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\docType{class}
\name{contacts-class}
\alias{contacts-class}
\alias{contacts}
\title{An S4 class for multi-contact information
from the Neotoma Paleoecology Database.}
\description{
An S4 class for multi-contact information
from the Neotoma Paleoecology Database.
}
