% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{coordinates,sites-method}
\alias{coordinates,sites-method}
\title{Return the latitude and longitude of sites}
\usage{
\S4method{coordinates}{sites}(obj, ...)
}
\arguments{
\item{obj}{A sites object}

\item{...}{Additional parameters associated with the call.}
}
\value{
\code{data.frame} object with site coordinates.
}
\description{
Return the latitude and longitude of sites
}
