% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{coordinates}
\alias{coordinates}
\title{Obtain coordinates from a sites object.}
\usage{
coordinates(obj, ...)
}
\arguments{
\item{obj}{A sites object}

\item{...}{Additional parameters associated with the call.}
}
\value{
dataframe with coordinate values
}
\description{
Obtain coordinates from a sites object.
}
