% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{count}
\alias{count}
\title{Count the number of sites, datasets, or collectionunits associated with a site object.}
\usage{
count(x, level = "sites")
}
\arguments{
\item{level}{The level at which you want to count. Can be sites, collunits, or datasets. Default is sites.}
}
\value{
A count of the specified level in the object
}
\description{
Count the number of sites, datasets, or collectionunits associated with a site object.
}
