% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\title{S4 class for dataset information}
\description{
The standard object class for datasets from the Neotoma Paleoecology Database.
}
