% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\title{S4 class for dataset information}
\value{
object of class \code{dataset}
}
\description{
The standard object class for datasets
from the Neotoma Paleoecology Database.
}
