% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{datasets-class}
\alias{datasets-class}
\title{S4 class for datasets information}
\value{
object of class \code{datasets}
}
\description{
The grouped class for datasets from
the Neotoma Paleoecology Database.
}
