% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{datasets,collunit-method}
\alias{datasets,collunit-method}
\title{Extract \code{datasets} from a \code{collunit} object.}
\usage{
\S4method{datasets}{collunit}(object)
}
\arguments{
\item{object}{A \code{collunit} object}
}
\value{
\code{datasets} from a \code{collunit} object
}
\description{
Extract \code{datasets} from a \code{collunit} object.
}
