% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-methods.R
\name{datasets,collunits-method}
\alias{datasets,collunits-method}
\title{Extract \code{datasets} from a \code{collunits} object.}
\usage{
\S4method{datasets}{collunits}(object)
}
\arguments{
\item{object}{A \code{collunits} object}
}
\value{
\code{datasets} from a \code{collunits} object
}
\description{
Extract \code{datasets} from a \code{collunits} object.
}
