% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_subset-methods.R, R/subset-methods.R
\name{datasets,site-method}
\alias{datasets,site-method}
\title{Extract datasets from a sites object.}
\usage{
\S4method{datasets}{site}(object)

\S4method{datasets}{site}(object)
}
\arguments{
\item{object}{A sites object}
}
\description{
Extract datasets from a sites object.

Extract datasets from a sites object.
}
