% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/subset-methods.R
\name{datasets}
\alias{datasets}
\alias{datasets,collunit-method}
\alias{datasets,sites-method}
\alias{datasets,site-method}
\title{Extract datasets from a sites object.}
\usage{
datasets(object)

\S4method{datasets}{collunit}(object)

\S4method{datasets}{sites}(object)

\S4method{datasets}{site}(object)
}
\arguments{
\item{object}{A sites object}
}
\value{
datasets object specific to the metadata contained in datasets
}
\description{
If the sites object contains datasets, then the datasets
will be returned. If the sites object does not contain datasets then
the user can apply \code{get_datasets()} to the object.
}
