% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{datasets}
\alias{datasets}
\title{Extract datasets from a sites object.}
\usage{
datasets(object)
}
\arguments{
\item{object}{A sites object}
}
\description{
Extract datasets from a sites object.
}
