% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{datasets}
\alias{datasets}
\title{Extract datasets from a sites object.}
\usage{
datasets(object)
}
\arguments{
\item{object}{A sites object}
}
\value{
datasets object specific to the metadata contained in datasets
}
\description{
If the sites object contains datasets, then the datasets
will be returned. If the sites object does not contain datasets then
the user can apply \code{get_datasets()} to the object.
}
