% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{doi,site-method}
\alias{doi,site-method}
\title{Obtain dataset DOIs from records.}
\usage{
\S4method{doi}{site}(x)
}
\arguments{
\item{x}{a neotoma2 \code{site} object}
}
\description{
Given complete dataset objects in Neotoma (must have used
\code{get_datasets()} or \code{get_downloads()}), return the dataset
doi for the record.
}
\examples{
mb <- get_sites(gpid = "Manitoba", datasettype = "pollen")
mb_ds <- get_datasets(mb)
doi(mb_ds)
}
