% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R,
%   R/publication-methods.R, R/site-methods.R
\name{doi}
\alias{doi}
\alias{doi,publication-method}
\alias{doi,site-method}
\title{Obtain the DOI for publications or datasets.}
\usage{
doi(x)

\S4method{doi}{publication}(x)

\S4method{doi}{site}(x)
}
\arguments{
\item{x}{Object with DOIs associated to it.}
}
\value{
doi object with DOI information
}
\description{
Obtain the DOI for publications or datasets.
}
