% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{doi}
\alias{doi}
\title{Obtain the DOI for publications or datasets.}
\usage{
doi(x)
}
\arguments{
\item{x}{Object with DOIs associated to it.}
}
\value{
doi object with DOI information
}
\description{
Obtain the DOI for publications or datasets.
}
