% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
filter(x, ...)
}
\arguments{
\item{x}{A site, dataset or download.}

\item{...}{arguments to filter by: latmin, latmax
longmin, longmax, elevmin, elevmax, datasettype.}
}
\description{
filter
}
