% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
filter(
  x,
  latmin = NA,
  latmax = NA,
  longmin = NA,
  longmax = NA,
  elevmin = NA,
  elevmax = NA,
  type = NA,
  ...
)
}
\arguments{
\item{x}{A site, dataset or download.}

\item{latmin}{Minimum latitude to filter by}

\item{latmax}{Maximum latitude to filter by}

\item{longmin}{Minimum longitude to filter by}

\item{elevmin}{Minimum elevation to filter by}

\item{elevmax}{Maximum elevation to filter by}

\item{type}{Type to filter by}

\item{...}{optional arguments to pass into \code{get_dataset}.}
}
\description{
filter
}
