% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_nulls.r
\name{fix_null}
\alias{fix_null}
\title{Fix null values from API responses}
\usage{
fix_null(x)
}
\arguments{
\item{x}{item that has null objects.}
}
\value{
\code{list} with \code{NA}s instead of \code{null} objects
}
\description{
API responses from the Neotoma API return many fields
as NULL values. This function turns the NULL values into logical NAs.
}
