% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.default}
\alias{get_datasets.default}
\title{Get Dataset Default}
\usage{
\method{get_datasets}{default}(x, ...)
}
\arguments{
\item{x}{Use a single number to extract dataset information}

\item{...}{accepted arguments, see details for more information.}
}
\value{
\code{sites} object with full metadata up to the \code{dataset} level
}
\description{
Get Dataset Default
}
