% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.default}
\alias{get_datasets.default}
\title{Get Dataset Default}
\usage{
\method{get_datasets}{default}(..., complete_data = FALSE)
}
\arguments{
\item{...}{contactid, datasettype,
altmin, altmax, loc, ageyoung, ageold, ageof}

\item{datasetid}{Use a single number to extract site information}
}
\description{
Get Dataset Default
}
