% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.numeric}
\alias{get_datasets.numeric}
\title{Get Dataset Numeric}
\usage{
\method{get_datasets}{numeric}(datasetid, ..., verbose = 0)
}
\arguments{
\item{x}{Use a single number to extract site information}
}
\description{
Get Dataset Numeric
}
