% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.numeric}
\alias{get_datasets.numeric}
\title{Get Dataset Numeric}
\usage{
\method{get_datasets}{numeric}(x, ...)
}
\arguments{
\item{x}{Use a single number to extract site information}

\item{...}{Additional parameters to get_datasets}
}
\description{
Get Dataset Numeric
}
\examples{
dontrun{
allds <- get_datasets(1:29)
plotLeaflet(allds)
}
}
