% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets - Data Sets Information for Fossil Sites}
\usage{
get_datasets(datasetid = NA, ...)
}
\arguments{
\item{datasetid}{integer An internal Neotoma dataset identifier}

\item{...}{arguments in ellipse form}
}
\description{
Get datasets - Data Sets Information for Fossil Sites
}
