% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets}
\usage{
get_datasets(datasetid = NA, ...)
}
\arguments{
\item{...}{accepted arguments: sites_object, contactid, datasettype, altmin, altmax, loc, ageyoung, ageold, ageof}

\item{x}{Use a single number to extract site information}
}
\value{
The function returns either a single item of class \code{"try-error"} describing
the reason for failure (either mis-defined parameters or an error from the Neotoma API),
or a table of sites, with rows corresponding to the number of individual sites and datasets
returned by the Neotoma API.
Each "site" object contains 6 parameters that can be accessed as well:
siteid, sitename, location, altitude, description, limited collection units information.
\item{ \code{siteid} }{site ID number}
\item{ \code{sitename} }{site's name}
\item{ \code{location} }{sf object that describes site's location}
\item{ \code{description} }{}
\item{ \code{collunits} }{limited information on collunits}
Each "collection unit" embedded in the "sites" object contains 6 parameters that can be accessed as well:
\item{ \code{collunitid}}{collection unit ID number}
\item{ \code{handle} }{collection unit's handle}
\item{ \code{collunitname} }{collection unit's name}
\item{ \code{colldate} }{date in collection unit}
\item{ \code{substrate} }{substrate}
\item{ \code{location} }{sf object that describes site's location}
\item{ \code{datasets} }{detailed information regarding dataset}
Each "dataset" nested in the "collection unit" contains the following detail of information:
\item{ \code{datasetid} }{dataset ID number}
\item{ \code{datasetname} }{site's name}
\item{ \code{datasettype} }{type of data found}
\item{ \code{location} }{sf object that describes site's location}
\item{ \code{notes} }{notes on the dataset}
\item{ \code{taxa table} }{taxa table}
\item{ \code{pi list} }{P.I. info}
\item{ \code{analyst} }{analyst info}
\item{ \code{metadata} }{dataset metadata}
}
\description{
Information for Fossil Datasets
Look for a dataset details using only a dataset ID or for multiple metadata.
Displays a table with the following columns: siteid, sitename, lat, long, and elev.
The function takes parameters defined by the user and returns a sites object
with more detailed information regarding datasets.
The user may define all or none of the possible fields.
The function contains data checks for each defined parameter.
}
\examples{
\dontrun{
To find all datasets with a min altitude of 12 and a max altitude of 25:
sites_12to25 <- get_datasets(altmin=12, altmax=25)

To find all datasets in Brazil
brazil <- '{"type": "Polygon", 
"coordinates": [[
 [-73.125, -9.102096738726443],
 [-56.953125,-33.137551192346145],
 [-36.5625,-7.710991655433217],
 [-68.203125,13.923403897723347],
 [-73.125,-9.102096738726443]
]]}'
brazil_datasets <- get_datasets(loc = brazil[1])
}
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
