% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.sites}
\alias{get_datasets.sites}
\title{Get Dataset Sites}
\usage{
\method{get_datasets}{sites}(sites_o, verbose = 0)
}
\arguments{
\item{datasetid}{Use a single number to extract site information}
}
\description{
Get Dataset Sites
}
