% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.sites}
\alias{get_datasets.sites}
\title{Get Dataset from a \code{sites} object.}
\usage{
\method{get_datasets}{sites}(sites, verbose = 0)
}
\description{
Get Dataset from a \code{sites} object.
}
