% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets.sites}
\alias{get_datasets.sites}
\title{Get Dataset from a \code{sites} object.}
\usage{
\method{get_datasets}{sites}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sites}.}

\item{...}{accepted arguments: sites_object, contactid, datasettype,}
}
\description{
Get Dataset from a \code{sites} object.
}
