% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documentation.R
\name{get_documentation}
\alias{get_documentation}
\title{get_documentation}
\usage{
get_documentation()
}
\description{
Open up the Neotoma R homepage.
}
\examples{
\donttest{
if (interactive()) {
 get_documentation()
}
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
