% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads.numeric}
\alias{get_downloads.numeric}
\title{get_downloads}
\usage{
\method{get_downloads}{numeric}(x, ..., complete_data = FALSE)
}
\arguments{
\item{...}{arguments in ellipse form}

\item{datasetid}{Use a single number to extract site information}
}
\description{
get_downloads
}
