% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads.numeric}
\alias{get_downloads.numeric}
\title{get_downloads}
\usage{
\method{get_downloads}{numeric}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Use a single number to extract site information}

\item{...}{arguments in ellipse form}
}
\description{
get_downloads
}
