% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads}
\alias{get_downloads}
\title{Get downloads - Data}
\usage{
get_downloads(datasetid = NA, ...)
}
\arguments{
\item{datasetid}{integer A collection unit ID}

\item{...}{arguments in ellipse form}
}
\description{
Get downloads - Data
}
