% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads.R
\name{get_downloads.sites}
\alias{get_downloads.sites}
\title{get_downloads sites}
\usage{
\method{get_downloads}{sites}(sites_o, ..., complete_data = FALSE)
}
\arguments{
\item{sites_o}{sites object}

\item{...}{arguments in ellipse form}
}
\description{
get_downloads sites
}
