% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications.default}
\alias{get_publications.default}
\title{Get publication information from Neotoma}
\usage{
\method{get_publications}{default}(...)
}
\description{
Get publication information from Neotoma
}
