% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications.numeric}
\alias{get_publications.numeric}
\title{Get publications using their unique identifier.}
\usage{
\method{get_publications}{numeric}(x, ...)
}
\arguments{
\item{x}{integer A contact ID}

\item{...}{\code{publicationid}
The unique numeric identifier associated with a publication in Neotoma.
\code{datasetid}
A unique identifier for a Neotoma dataset that is associated
with a publication.
\code{familyname} The full or partial last name for an individual author.
\code{pubtype} The publication type, from \code{get_tables("publicationtypes")}.
\code{year} The year the publication was released.
\code{search} A plain text search string used to search the citation.}
}
\value{
\code{publications} object
}
\description{
Get publications using their unique identifier.
}
\examples{
{
# We want the paper identified in Neotoma as 666:
get_publications(666)
}
}
