% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publications.R
\name{get_publications.publication}
\alias{get_publications.publication}
\title{Get publication information for Neotoma records}
\usage{
\method{get_publications}{publication}(x, ...)
}
\arguments{
\item{x}{A publication ID
The unique numeric identifier associated with a publication in Neotoma.}

\item{...}{Other options to pass to \code{get_publications()}
datasetid
A unique identifier for a Neotoma dataset that is associated
with a publication.
familyname The full or partial last name for an individual author.
pubtype The publication type, from \code{get_tables("publicationtypes")}.
year The year the publication was released.
search A plain text search string used to search the citation.}
}
\description{
Get publication information for Neotoma records
}
