% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.default}
\alias{get_sites.default}
\title{get_sites}
\usage{
\method{get_sites}{default}(...)
}
\arguments{
\item{...}{One of a set of possible query parameters discussed in details.}
}
\description{
get_sites
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
