% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.default}
\alias{get_sites.default}
\title{Get Site Information for Fossil Sites}
\usage{
\method{get_sites}{default}(..., verbose = 0)
}
\arguments{
\item{...}{accepted arguments: siteid, sitename, altmax, altmin, loc}
}
\description{
Get Site Information for Fossil Sites
}
\examples{
get_sites(sitename = "Alexander Lake")
}
