% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.numeric}
\alias{get_sites.numeric}
\title{Get Site from a site ID}
\usage{
\method{get_sites}{numeric}(x, ...)
}
\arguments{
\item{...}{arguments in ellipse form}

\item{siteid}{The numeric site ID from Neotoma}
}
\description{
Get Site from a site ID
}
\examples{
get_sites(siteid = 24)
}
