% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.numeric}
\alias{get_sites.numeric}
\title{Get Site Information for Fossil Sites}
\usage{
\method{get_sites}{numeric}(x, ..., verbose = 0)
}
\arguments{
\item{siteid}{The numeric site ID from Neotoma}
}
\description{
Get Site Information for Fossil Sites
}
