% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.numeric}
\alias{get_sites.numeric}
\title{Get Site Numeric}
\usage{
\method{get_sites}{numeric}(siteid, ...)
}
\arguments{
\item{x}{Use a single number to extract site information}
}
\description{
Get Site Numeric
}
