% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites.numeric}
\alias{get_sites.numeric}
\title{Get Site Information for Fossil Sites}
\usage{
\method{get_sites}{numeric}(x, ...)
}
\arguments{
\item{x}{The numeric site ID from Neotoma}

\item{...}{accepted arguments if numeric all_data}
}
\description{
Get Site Information for Fossil Sites
}
