% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Get Site Information for Fossil Sites}
\usage{
get_sites(siteid = NA, ...)
}
\arguments{
\item{siteid}{integer A unique (integer) site identifier from Neotoma.}

\item{contactname}{A full or partial name for an individual contributor to the database.}

\item{familyname}{The full or partial last name for an individual contributor to the database.}
}
\description{
Get Site Information for Fossil Sites
}
