% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Get Sites}
\usage{
get_sites(siteid = NA, ...)
}
\arguments{
\item{siteid}{Use a single number to extract site information}

\item{...}{accepted arguments: sitename, altmax (The coordinates to create an sf object), altmin (The coordinates to create an sf object)}
}
\description{
Get Sites
}
\details{
Information for Fossil Sites
Look for a site details using only a site ID or for multiple sites using possible sitenames, max latitudes or min latitudes.
Displays a table with the following columns siteid, sitename, lat, long, and elev.
Returns an object "sites" that contains multiple single "site" objects. Each of this objects has 6 parameters that can be accessed as well.
}
\examples{
get_sites(siteid = 24)
}
