% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens.default}
\alias{get_specimens.default}
\title{Get Specimen datasetid}
\usage{
\method{get_specimens}{default}(...)
}
\arguments{
\item{...}{Pass argument datasetid and the corresponding datasetid}
}
\value{
The function returns a specimens list
}
\description{
Get Specimen datasetid
}
\examples{
{
# To find specimens in datasetid 41610
my_specimens <- get_specimens(datasetid = 41610)
}
}
