% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimens.R
\name{get_specimens}
\alias{get_specimens}
\title{get_specimens}
\usage{
get_specimens(x = NA, ...)
}
\arguments{
\item{x}{Use a single specimenid}

\item{...}{passing datasetids}
}
\value{
The function returns a specimens list
}
\description{
Information for Specimens
}
\examples{
\dontrun{
# To find all datasets with a min altitude of 12 and a max altitude of 25:
my_specimens <- get_specimens(19832)
}
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
