% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speleothems.R
\name{get_speleothems}
\alias{get_speleothems}
\title{get_speleothems}
\usage{
get_speleothems(x = NA, ...)
}
\arguments{
\item{x}{A sites object.}
}
\description{
get_speleothems
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
