% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{get_stats}
\usage{
get_stats(start, end)
}
\arguments{
\item{start}{The starting month (from present == 0) for which to generate the summary.}

\item{end}{The ending month (from present == 0) for which to generate the summary.}
}
\value{
\code{data.frame} with summary statistics
}
\description{
Returns the count of sites, datasets, publications and other
objects added to Neotoma during the requested time period.
}
\author{
Socorro Dominguez \email{s.dominguez@ht-data.com}

Simon Goring \email{goring@wisc.edu}
}
