% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{get_taxa}
\usage{
get_taxa(x = NA, ...)
}
\arguments{
\item{x}{string taxa name or names}

\item{...}{accepted arguments, see details for more information.}
}
\value{
A Neotoma2 sites object with datasets with the requested taxa.
}
\description{
a sites object with the requested taxa.
}
