% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon.R
\name{get_taxon.default}
\alias{get_taxon.default}
\title{Get Taxa Default}
\usage{
\method{get_taxon}{default}(x, ...)
}
\arguments{
\item{x}{Use a taxon ID to extract site information}

\item{...}{accepted arguments, see details for more information.}
}
\value{
\code{sites} object containing the requested \code{taxa}
}
\description{
Get Taxa Default
}
