% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R, R/get_taxon.R
\name{get_taxon}
\alias{get_taxon}
\title{get_taxa}
\usage{
get_taxon(x = NA, ...)

get_taxon(x = NA, ...)
}
\arguments{
\item{x}{string taxa name or names}
}
\value{
A Neotoma2 sites object with datasets with the requested taxa.

A Neotoma2 sites object with datasets with the requested taxa.
}
\description{
a sites object with the requested taxa.

a sites object with the requested taxa.
}
