% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.collunits}
\alias{getids.collunits}
\title{Get object IDs from collectionunits.}
\usage{
\method{getids}{collunits}(x, order = TRUE)
}
\arguments{
\item{x}{A Neotoma2 \code{collunits} object.}

\item{order}{sort items by siteid, collunitid, datasetid}
}
\value{
\code{data.frame} containing \code{siteid}, \code{datasetid}, and \code{collunitid}
}
\description{
From a set of \code{collectionunit} objects, return the
collectionunit and dataset ids.
}
\examples{
{
marion <- get_sites(sitename = "Marion Lake")
collunitids <- getids(collunits(marion))
}
}
