% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids}
\alias{getids}
\title{Get object IDs}
\usage{
getids(x, order = TRUE)
}
\arguments{
\item{x}{A neotoma2 \code{sites} object.}
}
\description{
This function parses a site object, from \code{site} to
\code{dataset} level and returns a \code{data.frame} that contains the
site, collectionunit and dataset IDs for each element within the site.
}
