% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getids.r
\name{getids.site}
\alias{getids.site}
\title{Get object IDs from a site object.}
\usage{
\method{getids}{site}(x, order = TRUE)
}
\arguments{
\item{x}{A Neotoma2 \code{site} object.}

\item{order}{sort items by siteid, collunitid, datasetid}
}
\value{
\code{data.frame} containing \code{siteid}, \code{datasetid}, and \code{collunitid}
}
\description{
Get object IDs from a site object.
}
