% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R
\name{hash,collunit-method}
\alias{hash,collunit-method}
\title{Hash a collunit object}
\usage{
\S4method{hash}{collunit}(x)

\S4method{hash}{collunit}(x)
}
\arguments{
\item{object}{dataset object}
}
\description{
Hash a Neotoma collection unit object to add as an attribute
returning the hashed siteid, sitename, latitude, longitude and altitude.

Hash a Neotoma dataset object to add as an attribute
}
\examples{
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunit[[1]])
some_site <- get_sites(sitename = "Site\%")
hash(some_site[[1]]$collunits[[1]]$datasets[[1]])
}
