% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{length,datasets-method}
\alias{length,datasets-method}
\title{Length Method datasets}
\usage{
\S4method{length}{datasets}(x)
}
\arguments{
\item{x}{datasets object}
}
\value{
\code{int} that showcases the length of a \code{datasets} object
}
\description{
Length Method datasets
}
