% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{length,publications-method}
\alias{length,publications-method}
\title{Get the number of publications in a publications object.}
\usage{
\S4method{length}{publications}(x)
}
\arguments{
\item{x}{A publications object.}
}
\value{
\code{int} of the length of the \code{publications} object
}
\description{
Get the number of publications in a publications object.
}
