% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{names,publications-method}
\alias{names,publications-method}
\title{Get slot names for a publication object.}
\usage{
\S4method{names}{publications}(x)
}
\arguments{
\item{x}{A \code{publications} object.}
}
\value{
\code{string} with \code{publications} slots' names
}
\description{
Get slot names for a publication object.
}
