% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{pager}
\alias{pager}
\title{pager}
\usage{
pager(response, complete_response = FALSE, ...)
}
\arguments{
\item{x}{The result of the API call}

\item{counter}{If TRUE, modify the response to contain all datapoints}
}
\description{
An internal helper function to count through all offsets/limits
}
