% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_location.R
\name{parse_location}
\alias{parse_location}
\title{parse_location}
\usage{
parse_location(x)
}
\arguments{
\item{x}{location object}
}
\value{
\code{sf} object to parse as location.
}
\description{
Retrieve location from datasets WTK, GeoJson, bounding box
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
