% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseURL.R
\name{parseURL}
\alias{parseURL}
\title{Make the HTTP request}
\usage{
parseURL(x, dev = FALSE, ...)
}
\arguments{
\item{x}{The HTTP path for the particular API call.}

\item{dev}{Should the development API be used (for experimental endpoints), default \code{FALSE}}

\item{...}{Any query parameters passed from the function calling \code{parseURL}.}
}
\description{
An internal helper function used to connect to the Neotoma API
in a standard manner, and to provide basic validation of any response.
}
