% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeaflet.R
\name{plotLeaflet,site-method}
\alias{plotLeaflet,site-method}
\title{plotLeaflet}
\usage{
\S4method{plotLeaflet}{site}(object)
}
\arguments{
\item{object}{Site object to plot}
}
\value{
\code{leaflet} map
}
\description{
Plot a site on a leaflet map
}
\examples{
\donttest{
modernSites <- get_sites(keyword = "Modern")
plotLeaflet(modernSites[[1]])
}
}
