% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/plotLeaflet.R
\name{plotLeaflet}
\alias{plotLeaflet}
\alias{plotLeaflet,site-method}
\alias{plotLeaflet,ANY-method}
\title{plotLeaflet}
\usage{
plotLeaflet(object)

\S4method{plotLeaflet}{site}(object)

\S4method{plotLeaflet}{ANY}(object)
}
\arguments{
\item{object}{Sites object to plot}
}
\value{
leaflet map with site markers
}
\description{
Plot sites on a leaflet map
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
