% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLeaflet.R
\name{plotLeaflet}
\alias{plotLeaflet}
\title{plotLeaflet}
\usage{
plotLeaflet(object, save_im = FALSE, path = "")
}
\arguments{
\item{object}{Sites object to plot}

\item{save_im}{save output}

\item{path}{location where output should be saved in. save_im must be TRUE}
}
\description{
Plot sites on a leaflet map
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
