% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{plotLeaflet}
\alias{plotLeaflet}
\title{plotLeaflet}
\usage{
plotLeaflet(object)
}
\arguments{
\item{object}{Sites object to plot}
}
\value{
leaflet map with site markers
}
\description{
Plot sites on a leaflet map
}
\author{
Socorro Dominguez \email{s.dominguez@ht-data.com}
}
