% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\docType{class}
\name{publication-class}
\alias{publication-class}
\alias{publication}
\title{An S4 class for a single Neotoma publication.}
\description{
An S4 class for a single Neotoma publication.
}
