% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{publications-class}
\alias{publications-class}
\title{An S4 class for multi-publication information from the Neotoma
Paleoecology Database. This S4 class allows a single dataset to have one
or more publication classes associated with it.}
\value{
object of class \code{publications}
}
\description{
An S4 class for multi-publication information from the Neotoma
Paleoecology Database. This S4 class allows a single dataset to have one
or more publication classes associated with it.
}
