% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\docType{class}
\name{sample-class}
\alias{sample-class}
\alias{sample}
\title{S4 class for dataset information}
\description{
The standard object class for samples
from the Neotoma Paleoecology Database.
}
