% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{sample-class}
\alias{sample-class}
\title{S4 class for dataset information}
\value{
object of class \code{sample}
}
\description{
The standard object class for samples
from the Neotoma Paleoecology Database.
}
