% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{samples-class}
\alias{samples-class}
\title{S4 class for the set of samples}
\value{
object of class \code{samples}
}
\description{
The grouped class for samples from
the Neotoma Paleoecology Database.
}
