% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,collunit-method}
\alias{samples,collunit-method}
\title{samples}
\usage{
\S4method{samples}{collunit}(x)
}
\arguments{
\item{x}{collunit object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain elements from collunit
}
