% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,collunits-method}
\alias{samples,collunits-method}
\title{Get samples from a collectionunit or set of collection units:}
\usage{
\S4method{samples}{collunits}(x)
}
\arguments{
\item{x}{collunits object}
}
\description{
Obtain elements from collunits
}
\examples{
\dontrun{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
pollen <- samples(collunits(marion))
plot(value ~ I(-1 * age),
     data = pollen[pollen$variablename == "Cupressaceae",],
xlab = "Years before present",
ylab = "Cupressaceae pollen count")
}
}
