% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples,site-method}
\alias{samples,site-method}
\title{samples}
\usage{
\S4method{samples}{site}(x)
}
\arguments{
\item{x}{site object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain elements on the samples level
}
\examples{
\donttest{
marion <- get_sites(sitename = "Marion Lake") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_downloads()
pollen <- samples(marion)
}
}
