% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/samples.R
\name{samples}
\alias{samples}
\alias{samples,site-method}
\alias{samples,collunits-method}
\alias{samples,collunit-method}
\title{Obtain samples from a record or multiple records.}
\usage{
samples(x, chronname = NULL)

\S4method{samples}{site}(x, chronname = NULL)

\S4method{samples}{collunits}(x, chronname = NULL)

\S4method{samples}{collunit}(x, chronname = NULL)
}
\arguments{
\item{x}{sites object}

\item{chronname}{Optional character name of a chronology to report ages
against. When \code{NULL} (default) the Neotoma default-chronology precedence
is applied. If the named chronology is not present for a dataset, the
default is used and a warning is issued.}
}
\value{
data.frame with record information at sample level
}
\description{
Obtain samples from a record or multiple records.
}
