% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{samples}
\alias{samples}
\title{Obtain samples from a record or multiple records.}
\usage{
samples(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at sample level
}
\description{
Obtain samples from a record or multiple records.
}
