% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{selectMatch}
\alias{selectMatch}
\title{Select the best match for an object.}
\usage{
selectMatch(x, n)
}
\arguments{
\item{x}{object}

\item{n}{n elements that are a best match}
}
\description{
Select the best match for an object.
}
