% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{set_default}
\alias{set_default}
\title{Set the default chronology within a collectionunit.}
\usage{
set_default(x, n)
}
\arguments{
\item{x}{A chronologies object.}

\item{n}{The particular chronology to be used as the default.}
}
\value{
sites object with new default chronology
}
\description{
Set the default chronology within a collectionunit.
}
