% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sample.R
\name{set_sample}
\alias{set_sample}
\title{set Sample Information}
\usage{
set_sample(
  x = NA,
  ages = list(),
  igsn = NA_character_,
  datum = data.frame(),
  depth = NA_integer_,
  sampleid = NA_integer_,
  thickness = NA_integer_,
  samplename = NA_character_,
  sampleanalyst = list(),
  analysisunitid = NA_integer_,
  analysisunitname = NA_character_
)
}
\arguments{
\item{x}{Object to be set as a sample}

\item{ages}{ages}

\item{igsn}{IGSN character}

\item{datum}{dataframe of datum}

\item{depth}{integer representing depth}

\item{sampleid}{ID for sample}

\item{thickness}{thickness of core}

\item{samplename}{sample's name}

\item{sampleanalyst}{Analyst's contact name}

\item{analysisunitid}{Which analysis unit it is}

\item{analysisunitname}{Analysis Unit's name}
}
\value{
\code{sample} object
}
\description{
set Sample Information
}
\examples{
{
# Set an empty sample
my_sample <- set_sample()
}
}
