% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_site.R
\name{set_site}
\alias{set_site}
\title{set Site Information for Fossil Sites}
\usage{
set_site(
  siteid,
  sitename = NA_character_,
  coordinates = c(),
  description = NA_character_,
  notes = NA_character_,
  collunits = new("collunits"),
  altitude = NA_integer_
)
}
\arguments{
\item{siteid}{site unique identificator if available}

\item{sitename}{actual site name}

\item{coordinates}{coordinates to create sf object}

\item{description}{description of site}

\item{notes}{additional information of the site}

\item{collunits}{collection units in the site}

\item{altitude}{altitude/elevation of the site}
}
\description{
set Site Information for Fossil Sites
}
\examples{
my_site <- set_site(sitename="My Lake", coordinates = c(10, -30),
                    description = "my lake", altitude = 30)
}
