% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_speleothem.R
\name{set_speleothem}
\alias{set_speleothem}
\title{set Site Information for Fossil Sites}
\usage{
set_speleothem(
  x = NA,
  entityid = NA_integer_,
  entityname = NA_character_,
  siteid = NA_integer_,
  collectionunitid = NA_integer_,
  dripheight = NA_integer_,
  monitoring = FALSE,
  relativeage = NA_character_,
  speleothemtype = NA_character_,
  dripheightunits = NA_character_,
  entitycovertype = NA_character_,
  entrancedistance = NA_integer_,
  landusecovertype = NA_character_,
  speleothemdriptype = NA_character_,
  landusecoverpercent = NA_integer_,
  vegetationcovertype = NA_character_,
  entitycoverthickness = NA_integer_,
  entrancedistanceunits = NA_character_,
  vegetationcoverpercent = NA_integer_
)
}
\arguments{
\item{x}{Object to be set as a speleothem}

\item{siteid}{The unique site id for a site.  If this site is new to Neotoma then leave the ID as NA (the default).}

\item{sitename}{Actual site name as a character string.}

\item{geography}{An \code{sf} object representing the site location, either as a polygon or point.}

\item{altitude}{altitude/elevation of the site.}

\item{geopolitical}{The geopolitical unit in which the site is located.}

\item{area}{The area of the site or depositional basin in \emph{ha}.  Can be calculated from the polygon.}

\item{description}{Function to create new site objects for personal analysis. The new object will not be uploaded to the database.}

\item{notes}{additional information of the site}

\item{collunits}{Collection units in the site}
}
\value{
\code{site} object
}
\description{
set Site Information for Fossil Sites
}
