% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{show,collunit-method}
\alias{show,collunit-method}
\title{Show the collection unit information}
\usage{
\S4method{show}{collunit}(object)
}
\arguments{
\item{x}{collunit object}
}
\description{
Show the collection unit information
}
