% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{show,collunits-method}
\alias{show,collunits-method}
\title{Show the collection unit information}
\usage{
\S4method{show}{collunits}(object)
}
\arguments{
\item{object}{collunits object}
}
\value{
null used for side effects. Printing a data.frame
}
\description{
Show the collection unit information
}
