% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-methods.R
\name{show,contact-method}
\alias{show,contact-method}
\title{Show contact object}
\usage{
\S4method{show}{contact}(object)

\S4method{show}{contact}(object)
}
\arguments{
\item{object}{a \code{contact} object}
}
\value{
null - side effect for printing contact object

Null - prints a data.frame
}
\description{
Show contact object

Show a contact object
}
