% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{show,publication-method}
\alias{show,publication-method}
\title{Print publications to screen.}
\usage{
\S4method{show}{publication}(object)
}
\arguments{
\item{object}{A \code{publication} object.}
}
\value{
NULL - side effect function of printing a data.frame
}
\description{
Print publications to screen.
}
