% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma-methods.R
\name{show,sites-method}
\alias{show,sites-method}
\title{Show Sites objects as a dataframe}
\usage{
\S4method{show}{sites}(object)
}
\arguments{
\item{object}{sites object}
}
\description{
Show Sites objects as a dataframe
}
