% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{show,sites-method}
\alias{show,sites-method}
\title{Show sites objects as a dataframe}
\usage{
\S4method{show}{sites}(object)
}
\arguments{
\item{object}{sites object}
}
\value{
NULL - side effect for printing a \code{data.frame} object
}
\description{
Return a set of site objects as a single data.frame().
}
