% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R, R/collunits-methods.R,
%   R/contact-methods.R, R/dataset-methods.R, R/publication-methods.R,
%   R/site-methods.R, R/speleothem-methods.R, R/taxon-methods.R
\name{show,chronology-method}
\alias{show,chronology-method}
\alias{show,chronologies-method}
\alias{show,collunits-method}
\alias{show,collunit-method}
\alias{show,contacts-method}
\alias{show,contact-method}
\alias{show,dataset-method}
\alias{show,datasets-method}
\alias{show,publications-method}
\alias{show,publication-method}
\alias{show}
\alias{show,site-method}
\alias{show,sites-method}
\alias{show,speleothem-method}
\alias{show,speleothems-method}
\alias{show,taxon-method}
\alias{show,taxa-method}
\title{Display a \code{sites} object or nested slots.}
\usage{
\S4method{show}{chronology}(object)

\S4method{show}{chronologies}(object)

\S4method{show}{collunits}(object)

\S4method{show}{collunit}(object)

\S4method{show}{contacts}(object)

\S4method{show}{contact}(object)

\S4method{show}{dataset}(object)

\S4method{show}{datasets}(object)

\S4method{show}{publications}(object)

\S4method{show}{publication}(object)

\S4method{show}{site}(object)

\S4method{show}{sites}(object)

\S4method{show}{speleothem}(object)

\S4method{show}{speleothems}(object)

\S4method{show}{taxon}(object)

\S4method{show}{taxa}(object)
}
\arguments{
\item{object}{\code{sites}, \code{datasets}, \code{collunits}, \code{contacts} object}
}
\description{
Display a \code{sites} object or nested slots.
}
