% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-methods.R
\name{showMatch,publication-method}
\alias{showMatch,publication-method}
\title{Show matched publication objects.}
\usage{
\S4method{showMatch}{publication}(x)
}
\arguments{
\item{x}{A \code{publication} object.}
}
\value{
NULL printed matches with other publications
}
\description{
Show matched publication objects.
}
