% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{site-class}
\alias{site-class}
\title{An S4 class for site information}
\value{
object of class \code{site}
}
\description{
The standard object class for sites
from the Neotoma Paleoecology Database.
}
