% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{sites-class}
\alias{sites-class}
\title{An S4 class for multi-site information}
\description{
The standard object class for multi-sites
from the Neotoma Paleoecology Database.from
@returns object of class \code{sites}
}
