% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimens.R
\name{specimens,collunit-method}
\alias{specimens,collunit-method}
\title{specimens}
\usage{
\S4method{specimens}{collunit}(x)
}
\arguments{
\item{x}{\code{collunit} object}
}
\value{
\code{data.frame} with specimens summary table
}
\description{
Obtain specimen elements from a \code{collunit}
}
