% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{specimens}
\alias{specimens}
\title{Obtain specimens from a record or multiple records.}
\usage{
specimens(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information regarding specimens
}
\description{
Obtain specimens from a record or multiple records.
}
