% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothemdetails.R
\name{speleothemdetails,sites-method}
\alias{speleothemdetails,sites-method}
\title{speleothemdetails}
\usage{
\S4method{speleothemdetails}{sites}(x)
}
\arguments{
\item{x}{site object}
}
\value{
\code{data.frame} with speleothem records
}
\description{
Obtain elements on the speleothems level
Experimental function: API and behavior may change.
}
\examples{
\donttest{
tryCatch({
kesang <- get_sites(sitename = "Kesang cave") \%>\%
  get_datasets() \%>\%
  filter(datasettype == "pollen") \%>\%
  get_speleothems()
sp <- speleothemdetails(kesang)
}, error = function(e) {
message("Neotoma server not responding. Try again later.")
})
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
