% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothemdetails.R
\name{speleothemdetails,sites-method}
\alias{speleothemdetails,sites-method}
\title{speleothemdetails}
\usage{
\S4method{speleothemdetails}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain all speleothems within a sites object
}
\examples{
{
ds <- get_datasets(37302)
sp <- speleothems(ds)
}
}
