% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/speleothemdetails.R
\name{speleothemdetails}
\alias{speleothemdetails}
\alias{speleothemdetails,site-method}
\alias{speleothemdetails,collunits-method}
\alias{speleothemdetails,collunit-method}
\title{Obtain speleothems and samples from a record or multiple records.}
\usage{
speleothemdetails(x)

\S4method{speleothemdetails}{site}(x)

\S4method{speleothemdetails}{collunits}(x)

\S4method{speleothemdetails}{collunit}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at speleothem level
}
\description{
Obtain speleothems and samples from a record or multiple records.
}
