% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_classDefinitions.R
\docType{class}
\name{speleothems-class}
\alias{speleothems-class}
\title{S4 class for Speleothem information}
\value{
object of class \code{speleothem}
}
\description{
The grouped class for speleothems from
the Neotoma Paleoecology Database.
}
