% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speleothems.R
\name{speleothems,sites-method}
\alias{speleothems,sites-method}
\title{speleothems}
\usage{
\S4method{speleothems}{sites}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
\code{data.frame} with sample records
}
\description{
Obtain all speleothems within a sites object
Experimental function: API and behavior may change.
}
\examples{
{
tryCatch({
ds <- get_datasets(37302)
sp <- speleothems(ds)
}, error = function(e) {
message("Neotoma server not responding. Try again later.")
})
}
}
\author{
Socorro Dominguez \email{dominguezvid@wisc.edu}
}
