% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/speleothems.R
\name{speleothems}
\alias{speleothems}
\alias{speleothems,site-method}
\alias{speleothems,collunits-method}
\alias{speleothems,collunit-method}
\title{Obtain speleothems from a record or multiple records.}
\usage{
speleothems(x)

\S4method{speleothems}{site}(x)

\S4method{speleothems}{collunits}(x)

\S4method{speleothems}{collunit}(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at speleothem level
}
\description{
Obtain speleothems from a record or multiple records.
}
