% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{speleothems}
\alias{speleothems}
\title{Obtain speleothems from a record or multiple records.}
\usage{
speleothems(x)
}
\arguments{
\item{x}{sites object}
}
\value{
data.frame with record information at speleothem level
}
\description{
Obtain speleothems from a record or multiple records.
}
