% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[,site,numeric,ANY,ANY-method}
\alias{[,site,numeric,ANY,ANY-method}
\title{Get site field by numeric index}
\usage{
\S4method{[}{site,numeric,ANY,ANY}(x, i)
}
\arguments{
\item{x}{The site object}

\item{i}{The column indicator}
}
\description{
Get site field by numeric index
}
