% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-methods.R
\name{[[,datasets,numeric,ANY-method}
\alias{[[,datasets,numeric,ANY-method}
\title{Slicer}
\usage{
\S4method{[[}{datasets,numeric,ANY}(x, i)
}
\arguments{
\item{x}{datasets object}

\item{i}{iteration in datasets list}
}
\value{
sliced \code{dataset} object
}
\description{
Obtain one of the elements within a datasets list
}
