% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-methods.R
\name{[[,samples,numeric,ANY-method}
\alias{[[,samples,numeric,ANY-method}
\title{Slicer}
\usage{
\S4method{[[}{samples,numeric,ANY}(x, i)
}
\arguments{
\item{x}{samples object}

\item{i}{iteration in samples list}
}
\value{
\code{samples} sliced object
}
\description{
Obtain one of the elements within a samples list
}
