% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{[[,sites,numeric,ANY-method}
\alias{[[,sites,numeric,ANY-method}
\title{Slicer}
\usage{
\S4method{[[}{sites,numeric,ANY}(x, i)
}
\arguments{
\item{x}{sites object}

\item{i}{iteration in sites list}
}
\value{
sliced \code{site} object
}
\description{
Obtain one of the elements within a sites list
}
\examples{
{
some_site <- get_sites(sitename = "Site\%", limit=3)
some_site[[2]]
}
}
