% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{[[<-,collunits,ANY,ANY,ANY-method}
\alias{[[<-,collunits,ANY,ANY,ANY-method}
\title{Insert collunit}
\usage{
\S4method{[[}{collunits,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{collunits object}

\item{i}{iteration in collunits list}

\item{value}{The value to be used}
}
\value{
Modified \code{collunits}
}
\description{
Obtain one of the elements within a collunits list
}
