% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/publication-methods.R, R/samples-methods.R, R/site-methods.R,
%   R/speleothem-methods.R, R/taxon-methods.R
\name{[[<-,collunits-method}
\alias{[[<-,collunits-method}
\alias{sub-subset,collunits-method}
\alias{[[<-,collunits,ANY,ANY-method}
\alias{[[<-,datasets-method}
\alias{sub-subset,datasets-method}
\alias{[[<-,datasets,ANY,ANY-method}
\alias{[[<-,publications-method}
\alias{sub-subset,publications-method}
\alias{[[<-,publications,ANY,ANY-method}
\alias{[[<-,samples-method}
\alias{sub-subset,samples-method}
\alias{[[<-,samples,ANY,ANY-method}
\alias{sub-subset}
\alias{[[<-,sites-method}
\alias{[[<-,sites,ANY,ANY-method}
\alias{[[<-,speleothems-method}
\alias{sub-subset,speleothems-method}
\alias{[[<-,speleothems,ANY,ANY-method}
\alias{[[<-,taxa-method}
\alias{sub-subset,taxa-method}
\alias{[[<-,taxa,ANY,ANY-method}
\title{sub-subset}
\usage{
\S4method{[[}{collunits}(x, i) <- value

\S4method{[[}{datasets}(x, i) <- value

\S4method{[[}{publications}(x, i) <- value

\S4method{[[}{samples}(x, i) <- value

\S4method{[[}{sites}(x, i) <- value

\S4method{[[}{speleothems}(x, i) <- value

\S4method{[[}{taxa}(x, i) <- value
}
\arguments{
\item{x}{\code{neotoma2} object}

\item{i}{iteration in \code{neotoma2} object}

\item{value}{The value to be used}
}
\value{
\code{sites} object with reassigned values
}
\description{
Obtain one of the elements within a nested \code{neotoma2} object
}
