% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/publication-methods.R, R/samples-methods.R, R/site-methods.R,
%   R/speleothem-methods.R, R/taxon-methods.R
\name{[[<-,collunits,ANY,ANY,ANY-method}
\alias{[[<-,collunits,ANY,ANY,ANY-method}
\alias{[[<-,datasets,ANY,ANY,ANY-method}
\alias{[[<-,publications,ANY,ANY,ANY-method}
\alias{[[<-,samples,ANY,ANY,ANY-method}
\alias{sub-subset}
\alias{[[<-,speleothems,ANY,ANY,ANY-method}
\alias{[[<-,taxa,ANY,ANY,ANY-method}
\title{sub-subset}
\usage{
\S4method{[[}{collunits,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{datasets,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{publications,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{samples,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{sites,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{speleothems,ANY,ANY,ANY}(x, i) <- value

\S4method{[[}{taxa,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{neotoma2 object}

\item{i}{iteration in neotoma2 object}

\item{value}{The value to be used}
}
\value{
\code{sites} object with reassigned values
}
\description{
Obtain one of the elements within a nested neotoma2 object
}
