% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/site-methods.R, R/speleothem-methods.R, R/taxon-methods.R
\name{[,collunits,numeric-method}
\alias{[,collunits,numeric-method}
\alias{[,datasets,numeric-method}
\alias{[}
\alias{[,site,numeric-method}
\alias{[,site,character-method}
\alias{[,speleothems,numeric-method}
\alias{[,taxa,numeric-method}
\title{sub}
\usage{
\S4method{[}{collunits,numeric}(x, i)

\S4method{[}{datasets,numeric}(x, i)

\S4method{[}{sites,numeric}(x, i)

\S4method{[}{site,numeric}(x, i)

\S4method{[}{site,character}(x, i)

\S4method{[}{speleothems,numeric}(x, i)

\S4method{[}{taxa,numeric}(x, i)
}
\arguments{
\item{x}{The \code{sites} object}

\item{i}{The numeric index}
}
\value{
sliced \code{site} object
}
\description{
sub
}
