% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/site-methods.R, R/speleothem-methods.R, R/taxon-methods.R
\name{[,collunits,numeric,ANY-method}
\alias{[,collunits,numeric,ANY-method}
\alias{sub,collunits-method}
\alias{[,datasets,numeric,ANY-method}
\alias{sub,datasets-method}
\alias{[}
\alias{[,sites,numeric-method}
\alias{[,site,numeric,ANY-method}
\alias{[,site,numeric-method}
\alias{[,site,character,ANY-method}
\alias{[,site,character-method}
\alias{[,speleothems,numeric,ANY-method}
\alias{sub,speleothems-method}
\alias{[,taxa,numeric,ANY-method}
\alias{sub,taxa-method}
\title{sub}
\usage{
\S4method{[}{collunits,numeric,ANY}(x, i)

\S4method{[}{datasets,numeric,ANY}(x, i)

\S4method{[}{sites,numeric,ANY}(x, i)

\S4method{[}{site,numeric,ANY}(x, i)

\S4method{[}{site,character,ANY}(x, i)

\S4method{[}{speleothems,numeric,ANY}(x, i)

\S4method{[}{taxa,numeric,ANY}(x, i)
}
\arguments{
\item{x}{The \code{sites} object}

\item{i}{The numeric index}
}
\value{
sliced \code{site} object
}
\description{
sub
}
