% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{[<-,collunit,character,ANY,ANY-method}
\alias{[<-,collunit,character,ANY,ANY-method}
\title{Assign collunit field by numeric index}
\usage{
\S4method{[}{collunit,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{The collunit object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
sliced element
}
\description{
Assign collunit field by numeric index
}
