% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/samples-methods.R, R/site-methods.R, R/speleothem-methods.R,
%   R/taxon-methods.R
\name{[<-,collunit,character-method}
\alias{[<-,collunit,character-method}
\alias{subset,collunit-method}
\alias{[<-,collunit,numeric-method}
\alias{[<-,dataset,character-method}
\alias{subset,dataset-method}
\alias{[<-,dataset,numeric-method}
\alias{[<-,sample,character-method}
\alias{subset,sample-method}
\alias{subset}
\alias{[<-,site,character-method}
\alias{[<-,site,numeric-method}
\alias{[<-,speleothem,character-method}
\alias{subset,speleothem-method}
\alias{[<-,speleothem,numeric-method}
\alias{[<-,taxon,character-method}
\alias{subset,taxon-method}
\alias{[<-,taxon,numeric-method}
\title{subset}
\usage{
\S4method{[}{collunit,character}(x, i) <- value

\S4method{[}{collunit,numeric}(x, i) <- value

\S4method{[}{dataset,character}(x, i) <- value

\S4method{[}{dataset,numeric}(x, i) <- value

\S4method{[}{sample,character}(x, i) <- value

\S4method{[}{site,character}(x, i) <- value

\S4method{[}{site,numeric}(x, i) <- value

\S4method{[}{speleothem,character}(x, i) <- value

\S4method{[}{speleothem,numeric}(x, i) <- value

\S4method{[}{taxon,character}(x, i) <- value

\S4method{[}{taxon,numeric}(x, i) <- value
}
\arguments{
\item{x}{A \code{neotoma2} object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{site} object with reassigned character values
}
\description{
subset
}
