% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R, R/dataset-methods.R,
%   R/samples-methods.R, R/site-methods.R, R/speleothem-methods.R,
%   R/taxon-methods.R
\name{[<-,collunit,character,ANY-method}
\alias{[<-,collunit,character,ANY-method}
\alias{subset,collunit-method}
\alias{[<-,collunit,numeric,ANY-method}
\alias{[<-,dataset,character,ANY-method}
\alias{subset,dataset-method}
\alias{[<-,dataset,numeric,ANY-method}
\alias{[<-,sample,character,ANY-method}
\alias{subset,sample-method}
\alias{subset}
\alias{[<-,site,character-method}
\alias{[<-,site,character,ANY-method}
\alias{[<-,site,numeric,ANY-method}
\alias{[<-,site,numeric-method}
\alias{[<-,speleothem,character,ANY-method}
\alias{subset,speleothem-method}
\alias{[<-,speleothem,numeric,ANY-method}
\alias{[<-,taxon,character,ANY-method}
\alias{subset,taxon-method}
\alias{[<-,taxon,numeric,ANY-method}
\title{subset}
\usage{
\S4method{[}{collunit,character,ANY}(x, i) <- value

\S4method{[}{collunit,numeric,ANY}(x, i) <- value

\S4method{[}{dataset,character,ANY}(x, i) <- value

\S4method{[}{dataset,numeric,ANY}(x, i) <- value

\S4method{[}{sample,character,ANY}(x, i) <- value

\S4method{[}{site,character,ANY}(x, i) <- value

\S4method{[}{site,numeric,ANY}(x, i) <- value

\S4method{[}{speleothem,character,ANY}(x, i) <- value

\S4method{[}{speleothem,numeric,ANY}(x, i) <- value

\S4method{[}{taxon,character,ANY}(x, i) <- value

\S4method{[}{taxon,numeric,ANY}(x, i) <- value
}
\arguments{
\item{x}{A \code{neotoma2} object.}

\item{i}{The column indicator.}

\item{value}{The value to be used.}
}
\value{
\code{site} object with reassigned character values
}
\description{
subset
}
