% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{summary,sites-method}
\alias{summary,sites-method}
\title{Summary of objects within a sites object.}
\usage{
\S4method{summary}{sites}(object, ...)
}
\arguments{
\item{x}{sites object 1}
}
\description{
Summary of objects within a sites object.
}
