% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-methods.R
\name{summary,sites-method}
\alias{summary,sites-method}
\title{Summary of objects within a sites object.}
\usage{
\S4method{summary}{sites}(object, ...)
}
\arguments{
\item{object}{sites object}

\item{...}{additional properties passed to \code{summary}}
}
\value{
\code{data.frame} object with site summary information
}
\description{
This function summarizes a sites object, from \code{site} level
and returns a \code{data.frame} that contains the site ID, sitename,
collectionunit ID, count of chronologies, count of datasets
and types of datasets within the site.
}
