% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{taxa,sites-method}
\alias{taxa,sites-method}
\title{Extract taxonomic data from a set of sites.}
\usage{
\S4method{taxa}{sites}(object)
}
\arguments{
\item{object}{A \code{sites} object.}
}
\value{
A \code{data.frame} reporting the taxa/data objects, units, &cetera within the set of records.
}
\description{
From a sites object,
}
