% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R, R/taxa.R
\name{taxa}
\alias{taxa}
\alias{taxa,site-method}
\alias{taxa,collunits-method}
\alias{taxa,collunit-method}
\title{taxa}
\usage{
taxa(object)

\S4method{taxa}{site}(object)

\S4method{taxa}{collunits}(object)

\S4method{taxa}{collunit}(object)
}
\arguments{
\item{object}{Sites object to extract taxa table from}
}
\value{
data.frame with taxa records
}
\description{
Show the samples table
}
