% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{taxa}
\alias{taxa}
\title{taxa}
\usage{
taxa(object)
}
\arguments{
\item{object}{Sites object to extract taxa table from}
}
\value{
data.frame with taxa records
}
\description{
Show the samples table
}
