% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toJSON.R
\name{toJSON,sites-method}
\alias{toJSON,sites-method}
\title{toJSON}
\usage{
\S4method{toJSON}{sites}(x = NA)
}
\arguments{
\item{x}{sites R object to be converted}
}
\value{
The function returns a json object
}
\description{
Convert a Site object into a json file for API management
}
\examples{
\dontrun{
# To find all sites that contain the string "Alex\%"
alex.sites <- get_sites(sitename="Alex\%")
Convert the object to json
toJSON(alex.sites)}
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
