% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_genericDefinitions.R
\name{toJSON}
\alias{toJSON}
\title{toJSON}
\usage{
toJSON(x)
}
\arguments{
\item{x}{Sites object to extract taxa table from}
}
\value{
JSON translation of \code{sites} object to JSON
}
\description{
Export toJSON
}
\author{
Socorro Dominguez \email{sedv8808@gmail.com}
}
