% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWide.R
\name{toWide}
\alias{toWide}
\title{toWide}
\usage{
toWide(x, groupby = "age", operation = "prop")
}
\arguments{
\item{x}{dataframe object with samples}

\item{groupby}{User can groupby 'age' or 'depth'}

\item{operation}{label or vector of operations to be chosen from: 'prop', 'sum',
'count', 'presence'.}
}
\description{
Obtain a wide table with information regarding of samples grouped by
variablename and depth/age.
}
