% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology-methods.R
\name{write.csv,chronologies-method}
\alias{write.csv,chronologies-method}
\title{write CSV}
\usage{
\S4method{write.csv}{chronologies}(x, ...)
}
\arguments{
\item{x}{chronologies object}

\item{...}{Additional parameters associated with the call.}
}
\value{
null, called for side effects
}
\description{
write CSV
}
