% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collunits-methods.R
\name{write.csv,collunits-method}
\alias{write.csv,collunits-method}
\title{write CSV}
\usage{
\S4method{write.csv}{collunits}(x, ...)
}
\arguments{
\item{x}{collunits object}

\item{...}{Additional parameters associated with the call.}
}
\value{
null side effect for saving a CSV file.
}
\description{
write CSV
}
